% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_behaviour.R
\name{class_behaviour}
\alias{class_behaviour}
\alias{daedalus_old_behaviour}
\alias{daedalus_new_behaviour}
\alias{is_daedalus_behaviour}
\alias{print.daedalus_behaviour}
\title{Represent behaviour-change mechanisms in DAEDALUS}
\usage{
daedalus_old_behaviour(rate = 0.001, lower_limit = 0.2)

daedalus_new_behaviour(
  hospital_capacity,
  behav_effectiveness = 0.5,
  baseline_optimism = 0.5,
  responsiveness = 1.5,
  k0 = 4.59,
  k1 = -9.19
)

is_daedalus_behaviour(x)

\method{print}{daedalus_behaviour}(x, ...)
}
\arguments{
\item{rate}{The marginal rate of decrease in the scaling factor for each
additional daily death in the 'old' behavioural model.}

\item{lower_limit}{The lower limit of the scaling factor in the 'old'
behavioural model; prevents scaling factor from reaching zero.}

\item{hospital_capacity}{The emergency hospital capacity of a country, but
\strong{may also be} a \verb{<daedalus_country>} object from which the hospital
capacity is extracted. See \eqn{\bar H} in Details.}

\item{behav_effectiveness}{A single double value for the effectiveness of
adopting behavioural measures against the risk of infection. Expected to be
in the range \eqn{[0, 1]}, where 0 represents no effectiveness, and 1
represents full effectiveness. See \eqn{\delta} in Details. Defaults to
0.5.}

\item{baseline_optimism}{A single double value for the baseline optimism
about pandemic outcomes. Expected to be in the range \eqn{[0, 1]}.
See \eqn{\bar B} in Details. Defaults to 0.5.}

\item{responsiveness}{A single double value for the population responsiveness
to an epidemic signal. Must have a lower value of 0, but the upper bound is
open. See \eqn{k_2} in Details. Defaults to 1.5.}

\item{k0}{A single, optional double value which is a scaling parameter for
the sigmoidal relationship between \eqn{p_t} and \code{behav_effectiveness}.
See \eqn{k_0} in Details. Defaults to 4.59.}

\item{k1}{A single, optional double value which is another scaling parameter
for the sigmoidal relationship between \eqn{p_t} and \code{behav_effectiveness}.
See \eqn{k_1} in Details. Defaults to -9.19.}
}
\value{
An object of class \verb{<daedalus_behaviour>} which inherits from
\verb{<daedalus_response>}. This is primarily a list holding behavioural
parameters.
}
\description{
Helper functions to prepare \verb{<daedalus_behaviour>} class responses
corresponding to either old or new behavioural change mechanisms.
Inherits from \verb{<daedalus_response>} to allow for more control in future over
when these mechanisms are active.
}
\details{
Daedalus currently supports two behavioural models (and the option to have
neither model active). These models simulate population-level changes towards
adopting behaviours that help reduce the risk of infection, which might be
expected during a major disease outbreak.

In both models, the transmission rate of the infection \eqn{\beta} is
reduced for all transmissions by a scaling factor; the models differ in how
the scaling factor is calculated.

\strong{Note that} a major issue with including this in a model run (any value
other than \code{"off"}) is that it leads to substantially lower response costs,
and generally better health outcomes (lives lost), \strong{without accounting for
any attendant economic or social costs}. As such, please treat the
behavioural models as experimental.
\subsection{Deaths-based public concern model}{

This model existed prior to \verb{daedalus v0.2.25} and is referred to as the
\code{"old"} behavioural model.

The scaling factor for \eqn{\beta} in this model is
\deqn{
(1 - \text{rate})^{\Delta D} \times (1 - L) + L
}
where \eqn{\text{rate}} is the marginal rate of reduction for each additional
death per day (\eqn{\Delta D}), and \eqn{L} is the lower limit of the scaling
factor to prevent the FOI going to zero.
}

\subsection{Optimism-responsiveness-effectiveness model}{

This model is referred to as the \code{"new"} behavioural model, and the scaling
factor for \eqn{\beta} is given by

\deqn{
\mathcal{B}(p_t, \delta) = p_t(1 - \delta)[p_t(1 - \delta) + (1 - p_t)] +
(1 - p_t)[p_t(1 - \delta) + (1 - p_t)]
}

where \eqn{p_t} is the proportion of the population taking protective
behaviour and \eqn{\delta} is the effectiveness of the behaviour, both in the
range \eqn{[0, 1]}. When either is zero, \eqn{\beta} is not scaled down.

\eqn{p_t} is computed during the model run, and is given by

\deqn{
p_t = \frac{1}{1 + \text{exp}\left\{-\left(k_0 + k_1 \bar B +
k_2 \frac{H_t}{\bar H}\right)\right\}}
}

where \eqn{k_0 = 4.59, k_1 = -9.19, k_2} are scaling parameters.
\eqn{k_0, k_1} have constant values that cannot be changed by the user, and
these have been chosen to produce a sigmoidal relationship between
\eqn{\bar B} and \eqn{p_t}.

\eqn{k_2} a constant user-supplied parameter that should be interpreted as
population responsiveness to a signal of epidemic severity
\eqn{H_t / \bar H}, which is the relative burden on hospital capacity
(\eqn{H_t}: hospital demand at time \eqn{t}, and
\eqn{\bar H}: emergency hospital capacity).

\eqn{\bar B} is a constant user-supplied parameter that captures the baseline
population optimism about the outbreak.
}
}
