% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_country.R
\name{class_country}
\alias{class_country}
\alias{daedalus_country}
\alias{is_daedalus_country}
\alias{print.daedalus_country}
\title{Represent countries and territories for DAEDALUS}
\usage{
daedalus_country(country, contact_matrix = NULL, group_working_age = NULL)

is_daedalus_country(x)

\method{print}{daedalus_country}(x, ...)
}
\arguments{
\item{country}{A string giving the country or territory name, or ISO2 or
ISO3 code; must be from among \link[daedalus.data:country_codes_iso2c]{daedalus.data::country_codes_iso2c} or
\link[daedalus.data:country_codes_iso3c]{daedalus.data::country_codes_iso3c} or \link[daedalus.data:country_names]{daedalus.data::country_names}.}

\item{contact_matrix}{An optional named list of numeric matrices,
representing age-specific contact matrices in different contexts or settings.
Defaults to \code{NULL}, which selects the country-specific contact matrix
provided in \link[daedalus.data:country_data]{daedalus.data::country_data}.}

\item{group_working_age}{An optional value for the age-group that is
considered to be the working-age group. Defaults to \code{3}, which is taken from
an internal constant.}

\item{x}{An object of the \verb{<daedalus_country>} class.}

\item{...}{Other parameters passed to \code{\link[=print]{print()}}.}
}
\value{
\itemize{
\item \code{daedalus_country()} returns an object of the S3 class \verb{<daedalus_country>}
\item \code{is_daedalus_country()} returns a logical for whether an object is a
\verb{<daedalus_country>}.
\item \code{print.daedalus_country()} invisibly returns the \verb{<daedalus_country>}
object \code{x}. Called for printing side-effects.
}
}
\description{
Helper functions to create and work with S3 class
\verb{<daedalus_country>} objects for use with \code{\link[=daedalus]{daedalus()}}.
These objects store country parameters for reuse and have methods for easy
parameter access and editing, as well as processing raw country
characteristics for the DAEDALUS model.
}
\examples{
x <- daedalus_country("Canada")

x

# check whether `x` is a <country> object
is_daedalus_country(x)

# using assignment operators; must be assigned as a list
x$contact_matrix <- list(home = matrix(99, 4, 4), school = matrix(1, 4, 4))
x

}
