% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_daedalus.R
\name{class_daedalus}
\alias{class_daedalus}
\alias{as_daedalus}
\alias{validate_daedalus}
\alias{is_daedalus}
\alias{print.daedalus}
\title{Convert a list to \verb{<daedalus>} output}
\usage{
as_daedalus(x)

validate_daedalus(x)

is_daedalus(x)

\method{print}{daedalus}(x, ...)
}
\arguments{
\item{x}{An object of the \verb{<daedalus>} class.}
}
\value{
An object of the \verb{<daedalus>} output class.

Invisibly returns \code{x}; called primarily for its side effects of
erroring when the object does not satisfy the \verb{<daedalus>} class
requirements.

A logical for whether \code{x} is of the \verb{<daedalus>} class.

None; called for its printing side effects.
}
\description{
Convert a list to \verb{<daedalus>} output

Validate a potential \verb{<daedalus>} class object

Check if an object is of the \verb{<daedalus>} class

Print \verb{<daedalus>} class objects
}
\keyword{internal}
