% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_country.R
\name{country}
\alias{country}
\title{Represent countries and territories for DAEDALUS}
\usage{
country(
  name,
  parameters = list(contact_matrix = NULL, contacts_workplace = NULL,
    contacts_consumer_worker = NULL)
)
}
\arguments{
\item{name}{A string giving the country or territory name; must be from among
\link{country_names}.}

\item{parameters}{An optional named list of country parameters that are
allowed to be modified. Currently, users may only pass their own contact
matrix, workplace contacts, and consumer-worker contact matrix. If these are
not passed, default values are accessed from stored package data.}
}
\value{
An object of the S3 class \verb{<country>}.
}
\description{
Helper function to create custom S3 class \verb{<country>} objects
for use with \code{\link[=daedalus]{daedalus()}}. These objects store country parameters for reuse
and have methods for easy parameter access and editing, as well as
processing raw country characteristics for the DAEDALUS model.
}
\examples{
country("Canada")

country(
  "United Kingdom",
  parameters = list(contact_matrix = matrix(1, 4, 4))
)
}
