% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_helpers.R
\name{country_name_from_arg}
\alias{country_name_from_arg}
\title{Get country name from arg}
\usage{
country_name_from_arg(country)
}
\arguments{
\item{country}{A string which could be a country name, or an ISO2 or ISO3
code.}
}
\value{
A supported country name
}
\description{
Function to get a supported country name given an arg which
may be a country name or may be an ISO2 or ISO3 code. In all cases,
the function checks if the country identifier is supported.
}
\keyword{internal}
