% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_data.R
\docType{data}
\name{country_names}
\alias{country_names}
\title{Country names for DAEDALUS}
\format{
\subsection{\code{country_names}}{

A character vector of recognised country or territory names.
}
}
\usage{
country_names
}
\description{
Names for countries and territories included with the package.
}
\keyword{datasets}
