% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_data.R
\docType{data}
\name{country_names_codes}
\alias{country_names_codes}
\alias{country_names}
\alias{country_codes_iso3c}
\alias{country_codes_iso2c}
\title{Country names and ISO codes for DAEDALUS}
\format{
Character vectors of recognised country names, and 3-character or 2-character
codes.

An object of class \code{character} of length 67.

An object of class \code{character} of length 67.
}
\source{
Country codes are taken from the
\href{https://www.iso.org/iso-3166-country-codes.html}{ISO 3166 specification} and
generated from country names using \code{\link[countrycode:countrycode]{countrycode::countrycode()}}.
}
\usage{
country_names

country_codes_iso3c

country_codes_iso2c
}
\description{
Names and character codes for countries provided with the
package. Note that country names are commonly used names and may not be
official names.
\itemize{
\item \code{country_names}: Country names;
\item \code{country_codes_iso3c}: ISO 3166 alpha 3 (3 character) codes;
\item \code{country_codes_iso2c}: ISO 3166 alpha 2 (2 character) codes;
}
}
\examples{
country_names

country_codes_iso3c

country_codes_iso2c
}
\keyword{datasets}
