% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus-package.R
\docType{package}
\name{daedalus-package}
\alias{daedalus-package}
\title{daedalus: Model Health, Social, and Economic Costs of a Pandemic}
\description{
Model the health, education, and economic costs of directly transmitted respiratory virus pandemics, under different scenarios of prior vaccine investment, policy interventions, and public behavioural change, using the 'DAEDALUS' integrated health-economics model adapted from Haw et al. (2022) \doi{10.1038/s43588-022-00233-0}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jameel-institute/daedalus}
  \item \url{https://jameel-institute.github.io/daedalus/}
  \item Report bugs at \url{https://github.com/jameel-institute/daedalus/issues}
}

}
\author{
\strong{Maintainer}: Pratik Gupte \email{p.gupte24@imperial.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID})

Authors:
\itemize{
  \item Pratik Gupte \email{p.gupte24@imperial.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID})
  \item Patrick Doohan (\href{https://orcid.org/0000-0001-8076-1106}{ORCID})
  \item Robert Johnson (\href{https://orcid.org/0000-0002-7365-0042}{ORCID})
  \item Pablo Perez-Guzman (\href{https://orcid.org/0000-0002-5277-5196}{ORCID})
  \item Rich FitzJohn (\href{https://orcid.org/0000-0001-8888-3837}{ORCID})
  \item Emma Russell
  \item David Mears
  \item Katharina Hauck (\href{https://orcid.org/0000-0003-3138-4169}{ORCID})
}

Other contributors:
\itemize{
  \item Abdul Latif Jameel Institute for Disease and Emergency Analytics [funder]
  \item Imperial College of Science, Technology and Medicine [copyright holder, funder]
}

}
\keyword{internal}
