% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus2.R
\name{daedalus2}
\alias{daedalus2}
\title{DAEDALUS model implemented with dust}
\usage{
daedalus2(time_end = 100, ...)
}
\arguments{
\item{time_end}{An integer-like value for the number of timesteps
at which to return data. This is treated as the number of days with data
returned for each day. Defaults to 300 days.}

\item{...}{Other arguments to be passed to the ODE solver; these are passed
to \code{\link[deSolve:ode]{deSolve::ode()}}.}
}
\description{
\code{daedalus::daedalus2()} will eventually replace \code{daedalus::daedalus()} with
the model implemented in C++ to integrate with \pkg{dust2}
(and in future \pkg{dust}),
}
