% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus2.R
\name{daedalus2_internal}
\alias{daedalus2_internal}
\title{Internal function for daedalus2}
\usage{
daedalus2_internal(time_end, params, state)
}
\value{
A list of state values as returned by \code{dust2::dust_unpack_state()}.
}
\description{
Internal function for daedalus2
}
\keyword{internal}
