% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus.R
\name{daedalus_internal}
\alias{daedalus_internal}
\title{Internal function for daedalus}
\usage{
daedalus_internal(
  time_end,
  parameters,
  initial_state,
  flags,
  ode_control,
  n_groups
)
}
\value{
A list of state values as returned by \code{dust2::dust_unpack_state()}.
}
\description{
Internal function for daedalus
}
\keyword{internal}
