% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus_rtm.R
\name{daedalus_rtm}
\alias{daedalus_rtm}
\title{DAEDALUS Cpp model}
\usage{
daedalus_rtm(
  country,
  infection,
  response_strategy = NULL,
  response_time_start = 0,
  response_time_end = 0,
  hospital_capacity = NULL,
  initial_state_manual = list(p_infectious = 1e-07),
  auto_social_distancing = FALSE,
  social_distancing_mandate = 1,
  time_end = 300
)
}
\description{
DAEDALUS Cpp model
}
