% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{default_parameters}
\alias{default_parameters}
\title{Default parameters for DAEDALUS}
\usage{
default_parameters(...)
}
\arguments{
\item{...}{User-specified values to replace default parameter values.}
}
\description{
Useful defaults for the DAEDALUS model for use during
development.
}
\examples{
default_parameters(beta = 1.5 / 7.0)
}
