% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{.daedalus_ode}
\alias{.daedalus_ode}
\title{DAEDALUS system of ODEs}
\usage{
.daedalus_ode(t, state, parameters)
}
\arguments{
\item{t}{A single double value for the time.}

\item{state}{Vector of the initial state. The form of the vector should be
\eqn{X_i, for X \in S, E, I_s, I_a, R, H, D}, where values for each
of the compartments for each age group \eqn{i} are given consecutively.
May be a matrix to allow for multiple age groups. See \code{\link[=daedalus]{daedalus()}} for an
explanation of the epidemiological compartments.}

\item{parameters}{List for the parameters used in the simulation. See
\code{\link[=default_parameters]{default_parameters()}} for a list of parameters used in the user-facing model
function \code{\link[=daedalus]{daedalus()}}.}
}
\value{
A list with a single numeric vector of the same size as \code{state},
suitable as output for \pkg{deSolve} functions such as \code{\link[deSolve:lsoda]{deSolve::lsoda()}}.
}
\description{
A compartmental epidemiological model with four age groups. This
function is intended to be passed to the solvers from the \pkg{deSolve}
package. This is a work in progress and additional components will be added
in future.
}
\section{Details: DAEDALUS model}{
This section is a work in progress, and describes the DAEDALUS model in
brief.

The model is age-stratified, supports heterogeneity in social contacts
between age groups, and currently supports four age groups.
\subsection{Epidemiological model}{

The model has the following epidemiological compartments: susceptible
(\eqn{S}), exposed (\eqn{E}), infectious and symptomatic (\eqn{I_s}),
infectious asymptomatic (\eqn{I_a}), hospitalised (\eqn{H}), recovered
(\eqn{R}), and dead (\eqn{D}).

The model currently allows only uniform parameters for transitions between
compartments. Group specific parameters are expected to be supported in
future.
}
}

\keyword{internal}
