% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_npi.R
\name{dummy_npi}
\alias{dummy_npi}
\title{Dummy NPI}
\usage{
dummy_npi()
}
\value{
A \verb{<daedalus_npi>} object intended to have no effect; openness is
set to 1.0, and start time and duration are set to \code{NULL}.
}
\description{
Dummy NPI
}
\keyword{internal}
