% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_vaccination.R
\name{dummy_vaccination}
\alias{dummy_vaccination}
\title{Dummy vaccination}
\usage{
dummy_vaccination(country)
}
\value{
A \code{daedalus_vaccination} object intended to have no effect;
vaccination rate and efficacy are set to zero.
}
\description{
The efficacy of a dummy vaccination object is set to 50\% as a stop-gap
implementation of pre-existing population immunity. In scenarios where a
'true' vaccination scenario is to be passed, it doesn't matter. In scenarios
where no vaccination is intended, it allows any individuals with pre-existing
immunity to be only partially susceptible, while still preventing any
model-time vaccinations as the \code{rate} is set to 0.
}
\keyword{internal}
