% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{econ_constants}
\alias{econ_constants}
\alias{EDU_EFFECTIVENESS_REMOTE}
\alias{EDU_ANNUAL_ROR}
\alias{WORK_EXPECTED_YEARS}
\alias{EARNINGS_LOSS_DISCOUNT}
\title{Economic constants used in DAEDALUS}
\usage{
EDU_EFFECTIVENESS_REMOTE

EDU_ANNUAL_ROR

WORK_EXPECTED_YEARS

EARNINGS_LOSS_DISCOUNT
}
\description{
Economic constants used in DAEDALUS
}
\details{
The Daedalus model uses these constants for economic value calculations:
\enumerate{
\item \code{edu_effectiveness_remote}: The effectiveness of remote education.
\item \code{edu_annual_ror}: The rate of return on a year of education.
\item \code{work_expected_years}: The expected number of years of work for school-age
children.
\item \code{earnings_loss_discount}: The discounting rate used in the calculation of
lost earnings.
}
}
\keyword{econ_constant}
