% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closure_data.R
\docType{data}
\name{econ_sector_names}
\alias{econ_sector_names}
\title{Economic sector names}
\format{
\subsection{\code{econ_sector_names}}{

A character vector of length \code{N_ECON_SECTORS} (45).
}
}
\source{
Multiple sources; to be updated shortly. See processing details in
`data-raw/closure_data.R
}
\usage{
econ_sector_names
}
\description{
Names or descriptions of openness of economic sectors modelled
in \emph{daedalus}, to which closures apply.
}
\examples{
econ_sector_names
}
\keyword{datasets}
