% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{epi_constants}
\alias{epi_constants}
\alias{COMPARTMENTS}
\alias{idx_COMPARTMENTS}
\alias{N_EPI_COMPARTMENTS}
\alias{i_EPI_COMPARTMENTS}
\alias{N_MODEL_COMPARTMENTS}
\alias{N_INFECTION_SUBSYSTEM}
\title{Epidemiological compartments and indices}
\usage{
COMPARTMENTS

idx_COMPARTMENTS

N_EPI_COMPARTMENTS

i_EPI_COMPARTMENTS

N_MODEL_COMPARTMENTS

N_INFECTION_SUBSYSTEM
}
\value{
\code{COMPARTMENTS} returns a character vector of the epidemiological
compartment names. \code{idx_COMPARTMENTS} returns a list with the compartment
index.

All other constants return integer values.
}
\description{
Names and indices for the epidemiological compartments used in
DAEDALUS, for reuse in model code. The duplication is for ease of extracting
indices from a named list, and of extracting names without having to call
\code{names()}.
}
\details{
DAEDALUS has 8 epidemiological compartments: susceptible, exposed, infectious
and symptomatic ("infect_symp"), infectious and asymptomatic ("infect_asymp")
, hospitalised leading to recovery ("hospitalised_recov"),
hospitalised leading to death ("hospitalised_death"), recovered, and dead.

There are 2 additional compartments that track the number of new infections
and new hospitalisations.
}
\keyword{epi_constant}
\keyword{epi_constants}
