% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_helpers.R
\name{epi_output_helpers}
\alias{epi_output_helpers}
\alias{get_incidence}
\alias{get_epidemic_summary}
\alias{get_new_vaccinations}
\alias{get_attack_rate}
\title{Calculate daily incidences and summarise epidemic measures}
\usage{
get_incidence(
  data,
  measures = c("infections", "hospitalisations", "deaths"),
  groups = NULL
)

get_epidemic_summary(
  data,
  measures = c("infections", "hospitalisations", "deaths"),
  groups = NULL
)

get_new_vaccinations(data, groups = NULL)

get_attack_rate(
  data,
  measures = c("infections", "hospitalisations", "deaths"),
  groups = c("econ_sector", "age_group")
)
}
\arguments{
\item{data}{Either a \verb{<data.frame>} from a call to \code{get_data()} on a
\verb{<daedalus_output>} object, or such an object directly.}

\item{measures}{A character vector of one or more of the following, passed to
\code{get_incidence()} and \code{get_epidemic_summary()}:
\code{"infections"}, \code{"hospitalisations"} or \code{"deaths"} for the
measure to return. Defaults to returning all three in long format.

\code{get_daily_vaccinations()} does not accept a \code{measures} argument and only
provides the number of daily vaccinations.}

\item{groups}{An optional character vector of grouping variables that
correspond to model strata. Defaults to \code{NULL} which gives incidence across
the whole population. Allowed groups correspond to modelled strata:
\code{"age_group"}, \code{"vaccine_group"}, and \code{"econ_sector"}.

\code{get_attack_rate()} does not support \code{"vaccine_group"} due to a shifting
baseline effect.

\code{get_daily_vaccinations()} only accepts "\code{age_group}" and \code{"econ_sector"}.}
}
\value{
A \verb{<data.frame>} in long format, with one entry per
model timestep, measure, and group chosen.
\itemize{
\item \code{get_incidence()} returns a data frame with the number of daily new
infections, new hospitalisations, and/or new deaths in each of the groups
specified by \code{groups}.
\item \code{get_epidemic_summary()} returns a data frame with the total number of the
value specified in \code{measure} for each of the groups specified by \code{groups}.
\item \code{get_daily_vaccinations()} returns a data frame with columns for the
number of new daily vaccination in each combination of \code{groups} if provided.
Columns for the \code{groups} are added when \code{groups} are specified.
\item \code{get_attack_rate()} returns a data frame similar to
\code{get_epidemic_summary()}, but with a \code{p_affected} column giving the value as
a proportion of the initial group size (capped at 1.0).
}
}
\description{
Functions to quickly summarise timeseries data from
\code{\link[=daedalus]{daedalus()}} to provide daily values for infections, hospitalisations,
deaths, and vaccinations, while allowing grouping by different strata.
}
\examples{
data <- daedalus("Canada", "sars_cov_1")

# new infections
new_infections <- get_incidence(data, "infections")

# epidemic summary
get_epidemic_summary(
  data,
  groups = "age_group"
)

# get daily vaccinations
daily_vaccinations <- get_new_vaccinations(data)

# get attack rate
get_attack_rate(data, "infections", groups = "age_group")

}
