% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_data.R
\docType{data}
\name{epidemic_data}
\alias{epidemic_data}
\alias{infection_data}
\alias{epidemic_names}
\title{Infection characteristics for model epidemics}
\format{
\subsection{\code{infection_data}}{

A list with 7 elements, each corresponding to an epidemic (see \strong{Details}),
and providing 10 infection parameters:
\itemize{
\item \code{r0}: A single numeric value for the basic reproduction value of the
infection \eqn{R_0}.
\item \code{sigma}: A single numeric value > 0.0 for the rate of transition from the
exposed compartment to one of two infectious compartments.
\item \code{p_sigma}: A single numeric value in the range \eqn{(0.0, 1.0)} for the
proportion of infectious individuals who are also symptomatic. Asymptomatic
individuals can have a different contribution to the force of infection from
symptomatic individuals.
\item \code{epsilon}: A single numeric value for the relative contribution of
asymptomatic infectious individuals to the force of infection (compared to
symptomatic individuals).
\item \code{gamma_Is}: A single numeric value for the recovery rate of infectious
individuals who are not hospitalised.
\item \code{gamma_Ia}: A single numeric value for the recovery rate from asymptomatic
infection.
\item \code{gamma_H}: A numeric vector of length 4 for the age-specific recovery rate
for individuals who are hospitalised.
\item \code{eta}: A numeric vector of length \code{N_AGE_GROUPS} (4) for the age-specific
hospitalisation rate for individuals who are infectious and symptomatic.
\item \code{omega}: A numeric vector of length \code{N_AGE_GROUPS} (4) for the age-specific
mortality rate for individuals who are hospitalised.
\item \code{rho}: A single numeric value for the rate at which infection-derived
immunity wanes, returning individuals in the 'recovered' compartment to the
'susceptible' compartment.
}
}

\subsection{\code{epidemic_names}}{

A character vector with 7 elements.
}
}
\source{
See processing details in `data-raw/infection_data.R
}
\usage{
infection_data

epidemic_names
}
\description{
Infection parameters used in DAEDALUS to simulate epidemics of
interest.

\code{epidemic_names} provides the tags by which epidemics are known for users'
convenience.
}
\details{
Epidemics for which data are available are (pathogen in parentheses):
\itemize{
\item SARS 2004 (SARS-CoV-1)
\item Influenza 2009 (influenza A H1N1)
\item Influenza 1957 (influenza A H2N2)
\item Influenza 1918 (influenza A H1N1)
\item Covid-19 wild type (SARS-Cov-2 wild type)
\item Covid-19 Omicron (SARS-CoV-2 omicron)
\item Covid-19 Delta (SARS-CoV-2 delta).
}
}
\examples{
# check available epidemics
epidemic_names
}
\keyword{datasets}
