% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_size.R
\name{epidemic_summary}
\alias{epidemic_summary}
\alias{get_new_infections}
\alias{get_epidemic_summary}
\title{Get derived epidemic data}
\usage{
get_new_infections(data, grouping_vars = NULL)

get_epidemic_summary(
  data,
  measure = c("epidemic_size", "total_hospitalisations", "total_deaths"),
  grouping_vars = NULL
)
}
\arguments{
\item{data}{Either a \verb{<data.frame>} or a \verb{<daedalus_output>} object.}

\item{grouping_vars}{An optional character vector of grouping variables that
correspond to model strata. Defaults to \code{NULL}, with allowed values of
\code{age_group}, \code{vaccine_group}, and \code{econ_sector}.}

\item{measure}{A character vector of one or more of the following:
\code{"epidemic_size"}, \code{"total_hospitalisations"} or \code{"total_deaths"} for the
measure to return.}
}
\value{
A \verb{<data.frame>}.
\itemize{
\item \code{get_new_infections()} returns a data frame with the number of daily new
infections in each of the groups specified by \code{grouping_vars}.
\item \code{get_epidemic_summary()} returns a data frame with the measure specified
in \code{measure} for each of the groups specified by \code{grouping_vars}.
}
}
\description{
Functions to quickly summarise timeseries data from \code{daedalus()}
while allowing grouping by different strata.
}
\examples{
data <- daedalus("Canada", "sars_cov_1")

# new infections
new_infections <- get_new_infections(data)

# epidemic summary
get_epidemic_summary(
  data,
  measure = c("epidemic_size", "total_deaths", "total_hospitalisations"),
  grouping_vars = "age_group"
)

}
