% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reff_calculation.R
\name{get_beta}
\alias{get_beta}
\title{Calculate transmission parameter from infection and contact parameters}
\usage{
get_beta(infection, country)
}
\arguments{
\item{infection}{A \verb{<daedalus_infection>} object.}

\item{country}{A \verb{<daedalus_country>} object.}
}
\description{
Calculate transmission parameter from infection and contact parameters
}
\keyword{internal}
