% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_closure_info}
\alias{get_closure_info}
\title{Get closure time limits and calculate duration}
\usage{
get_closure_info(mutables)
}
\arguments{
\item{mutables}{An environment holding the mutable parameters.
See \code{\link[=prepare_mutable_parameters]{prepare_mutable_parameters()}} for a template.}
}
\value{
A three-element list of the start time, end time, and duration for which
closures are active.
}
\description{
Get closure time limits and calculate duration
}
\keyword{internal}
