% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus.R
\name{get_daedalus_multi_response_times}
\alias{get_daedalus_multi_response_times}
\title{Get response times from a dust2 system with multiple groups}
\usage{
get_daedalus_multi_response_times(output, n_groups)
}
\arguments{
\item{output}{dust2 output \code{daedalus_internal()}.}
}
\description{
Get response times from a dust2 system with multiple groups
}
\keyword{internal}
