% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus.R
\name{get_daedalus_response_times}
\alias{get_daedalus_response_times}
\title{Get model response times from dust2 output}
\usage{
get_daedalus_response_times(output)
}
\arguments{
\item{output}{dust2 output \code{daedalus_internal()}.}
}
\value{
A list of event start and end times, closure periods, and the
duration of each closure event, suitable for a \verb{<daedalus_output>} object.
}
\description{
Get model response times from dust2 output
}
\keyword{internal}
