% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_country.R, R/class_infection.R,
%   R/class_npi.R, R/class_output.R, R/class_vaccination.R, R/pkg_generics.R
\name{get_data}
\alias{get_data}
\alias{get_data.daedalus_country}
\alias{get_data.daedalus_infection}
\alias{get_data.daedalus_npi}
\alias{get_data.daedalus_output}
\alias{get_data.daedalus_vaccination}
\title{Get parameters from DAEDALUS classes}
\usage{
\method{get_data}{daedalus_country}(x, to_get, ...)

\method{get_data}{daedalus_infection}(x, to_get, ...)

\method{get_data}{daedalus_npi}(x, to_get, ...)

\method{get_data}{daedalus_output}(x, to_get = NULL, ...)

\method{get_data}{daedalus_vaccination}(x, to_get, ...)

get_data(x, ...)
}
\arguments{
\item{x}{An S3 class object from the \pkg{daedalus} package of the
\verb{<daedalus_country>}, \verb{<daedalus_infection>}, \verb{<daedalus_npi>},
\verb{<daedalus_output>}, and \verb{<daedalus_vaccination>} class.}

\item{to_get}{A string giving the name of the element of \code{x} to return.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Other arguments to class
methods. Class methods do not currently support any other arguments.}
}
\value{
Returns a member of \code{x}, with the class preserved
(e.g. numeric vector for a country \code{"demography"}).

For \verb{<daedalus_output>} objects, returns the model timeseries data when no
element is specified.
}
\description{
Generic and methods for S3 classes for safely getting class
parameters.
}
\examples{
# simple example of getting data
country_A <- daedalus_country("United Kingdom")
get_data(country_A, "demography")

get_data(country_A, "contact_matrix")

disease_x <- daedalus_infection("sars_cov_1", r0 = 1.9)
get_data(disease_x, "r0")

# get model data
output <- daedalus("Canada", "influenza_1918")
head(
  get_data(output)
)

}
\keyword{generics}
