% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{get_fiscal_costs}
\alias{get_fiscal_costs}
\title{Get pandemic fiscal costs from a model run}
\usage{
get_fiscal_costs(
  x,
  support_level = 0.2,
  price_vax = 1,
  price_npi = 1,
  uptake_npi = 1,
  interest_rate = 4,
  tax_rate = 35,
  spending_rate = 45,
  starting_debt = 0,
  productivity_loss_infection = 1
)
}
\arguments{
\item{x}{A \verb{<daedalus_output>} object from a call to \code{\link[=daedalus]{daedalus()}}, or a
\verb{<data.frame>} of an epidemic timeseries.

If \code{x} is a \verb{<data.frame>}, it must have the columns \code{"time"},
\code{"compartment"}, \code{"age_group"}, \code{"econ_sector"} and \code{"value"}, giving the
model time, the epidemiological compartment, and the number of individuals
of each age group and economic sector in each compartment at each time.
See \strong{Details} for more on the expectations around how this dataset is
organised.}

\item{support_level}{The proportion of pandemic-related economic losses that
a government compensates, as a proportion.}

\item{price_vax}{The per-dose price of vaccination.}

\item{price_npi}{The per-day and per-person price of implementing any
pandemic response. May include costs such as testing or masks.}

\item{uptake_npi}{The extent to which NPIs are taken up by the population;
essentially a number that modifies (reduces) the total cost of implementing
an NPI.}

\item{interest_rate}{The annual interest rate on government borrowing for
pandemic response.}

\item{tax_rate}{The annual mean tax rate on economic output; used to
calculate government revenue.}

\item{spending_rate}{The annual mean rate of public spending as a percentage
of GDP.}

\item{starting_debt}{The value of national starting debt. Currently assumed
to be zero while country-specific data are collected.}

\item{productivity_loss_infection}{A single number in the range
\eqn{[0, 1]} giving the loss in productivity associated with symptomatic
infection. Currently defaults to 1.0 for compatibility with earlier function
versions.}
}
\value{
A two-element list giving:
\itemize{
\item A list of \code{fiscal_costs} with elements giving the fiscal costs
\eqn{\text{TCG}_t} and a breakdown of these costs, as well as interest;
\item A list of \code{public_debt} of the public debt \eqn{b_t}, which is
the net of baseline public spending, pandemic response costs, and
pandemic-impacted revenue.
}
}
\description{
A helper function that post-processes a \verb{<daedalus_output>} object to
calculate the costs to a national government (fiscal costs) of responding to
a pandemic. Includes costs of economic support, vaccinations given, and NPIs
administered or implemented.
}
\details{
Default argument values are intended to be representative. This
function is intended to be called after \code{\link[=daedalus]{daedalus()}} and parameters
required to compute fiscal costs may need to be bundled along with model
outputs.
Note that all rates (interest rate, spending rate, and tax rate) are given as
annual percentages. Only the interest rate is converted to an daily value
from an annual one for use in compounding.
\subsection{Public spending}{

Spending is calculated as:

\deqn{G_t = \sigma(\text{GVA}^* - \text{GVA}_t) + Cv_t + Cp_t}

where \eqn{Cv_t} is the time-specific cost of vaccination, and is calculated
as the cost of new vaccinations in each timestep: \eqn{P_v \times \delta V_t}
.

\eqn{Cp_t} is the time-specific cost of implementing pandemic response,
and is calculated as \eqn{\psi (\bar N - D_t) P_p}, where
\eqn{\bar N - D_t} is the remaining number of individuals in the population,
and \eqn{\psi} is the proportion taking up any protection offered by the
response.
}

\subsection{Interest on spending and fiscal cost}{

We assume that the government borrows to spend on pandemic mitigation
measures outlined above such that the total cost to the public is then

\deqn{\text{TCG}_t = G_t + (1 + R_t^T) \text{TCG}_{t - 1}}

where \eqn{R_t^T} is the daily rate of interest to be paid on the borrowed
amount.

The interest rate is modelled as being constant over time. Users pass the
annual rate of interest as a percentage, and this is converted to a daily
rate using the internal function \code{annual_rate_daily()} as
\eqn{(1 + R_t^T)^{1 / 365} - 1}.
}

\subsection{Total public debt}{

The total public debt at the end of the pandemic \eqn{b_t} is then the sum of
:
\itemize{
\item total public spending on the pandemic \eqn{\text{TCG}_t},
\item existing day-to-day public spending \eqn{\bar G} which is assumed to be a
fraction of daily GDP \eqn{\nu \text{GDP}},
\item existing debt owed due to past daily spending (including on pandemic
mitigation), and interest to be paid on the debt,
\item less the revenues collected from taxation, \eqn{\mu \text{GVA}_{t-1}},
where \eqn{\mu} is the mean rate of taxation.
}

The daily GVA is the pre-pandemic GVA scaled by the available labour supply
during the pandemic, taking into account labour restrictions due to
illness-related absences and deaths, and response-related
restrictions.

GDP is calculated as the sum of sector-specific daily GVA, and existing debt
is currently assumed to be zero and is not included in the equation.

\deqn{b_t = \bar G + \text{TCG}_{t-1} + (1 + R_{t-1}^T) b_{t-1} -
  \mu \text{GVA}_{t-1}
}
}
}
\examples{
# get fiscal costs for UK with SARS-CoV-2
# and both closures and vaccinations
o <- daedalus(
  "GBR", "sars_cov_2_pre_alpha",
  "economic_closures", "high",
  time_end = 100
)
fc <- get_fiscal_costs(o)

# also works when no closures are applied
o <- daedalus(
  "CAN", "influenza_2009",
  time_end = 30
)

# Compare public debt added estimates with other estimates from Covid-19:
# https://www.ctf.ca/EN/EN/Newsletters/Perspectives/2020/3/200302.aspx
get_fiscal_costs(o)
}
