% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reff_calculation.R
\name{get_hospitalisations}
\alias{get_hospitalisations}
\title{Calculate total hospitalisations}
\usage{
get_hospitalisations(state)
}
\arguments{
\item{state}{The ODE state variable as an array. Must have at least \code{i_H}
columns.}
}
\description{
Calculate total hospitalisations
}
\keyword{internal}
