% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_helpers.R
\name{get_life_years_lost}
\alias{get_life_years_lost}
\title{Get life-years lost by demographic group.}
\usage{
get_life_years_lost(output, groups = c("none", "age_group"))
}
\arguments{
\item{output}{A \verb{<daedalu_output>} object.}

\item{groups}{Whether to get the life-years lost by age group. Selecting
\code{"none"} gives the total life-years lost.}
}
\value{
A \verb{<data.frame>} with the number of life-years lost, optionally per
age-group.
}
\description{
Get life-years lost by demographic group.
}
