% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helpers.R
\name{get_state_indices}
\alias{get_state_indices}
\title{Get a vector of state indices}
\usage{
get_state_indices(state_name, country)
}
\arguments{
\item{state_name}{The state name as a string.}

\item{country}{The country as a \verb{<daedalus_country>}. Needed to
correctly calculate the number of age and economic sector groups.}
}
\value{
A vector of numbers representing indices.
}
\description{
A helper function to get a vector of state indices for use in the
response sub-class constructors.
}
\keyword{internal}
