% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{get_value_lost_earnings}
\alias{get_value_lost_earnings}
\title{Calculate the present value of lost earnings due to educational disruption}
\usage{
get_value_lost_earnings()
}
\value{
A single number (around 20) that gives a coefficient of lost
earnings.
}
\description{
Calculate the present value of lost earnings due to educational disruption
}
\keyword{internal}
