% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus2.R
\name{initial_flags}
\alias{initial_flags}
\title{Initial values for model flags}
\usage{
initial_flags()
}
\value{
A vector of initial flag values; all flags are initially set to 'off'
or 0.0.
}
\description{
Initial values for model flags
}
\keyword{internal}
