% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{input_checkers}
\alias{input_checkers}
\alias{assert_class_fields}
\title{Check that a class has expected names}
\usage{
assert_class_fields(x, expected_fields)
}
\arguments{
\item{x}{An object to be tested as having \code{expected_fields} class member
names.}

\item{expected_fields}{A character vector of expected class member names.}
}
\value{
Invisibly returns \code{x}; primarily called for its side effects.
}
\description{
Check that a class has expected names
}
\keyword{internal}
