% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_country.R
\name{is_country}
\alias{is_country}
\title{Check if an object is a \verb{<country>}}
\usage{
is_country(x)
}
\arguments{
\item{x}{An object to be checked as being of the \verb{<country>} class.}
}
\value{
A single logical for whether \code{x} is a \verb{<country>}.
}
\description{
Check if an object is a \verb{<country>}
}
\examples{
is_country(country("Canada"))
}
