% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_country.R
\name{is_nulls_list}
\alias{is_nulls_list}
\title{Check for a list of NULLs}
\usage{
is_nulls_list(x)
}
\arguments{
\item{x}{A list to be tested}
}
\description{
Check for a list of NULLs
}
\keyword{internal}
