% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_value_data.R
\docType{data}
\name{life_value}
\alias{life_value}
\title{Values of statistical lives lost}
\format{
\subsection{\code{life_value}}{

A list with 180 elements, each corresponding to a country or territory, and
giving the age-group-specific value of a statistical life lost. Each list
element is a numeric vector of four values giving the value per age group,
in the following order: 0-4, 5-19, 20-65, 65+.
}
}
\source{
Multiple sources; see processing details in `data-raw/life_value.R.
}
\usage{
life_value
}
\description{
Values in dollars (scaled by purchasing power parity) of a
statistical life lost in each of the four DAEDALUS age groups, accounting
for life expectancy.
}
\keyword{datasets}
