% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closure.R
\name{make_closure_event}
\alias{make_closure_event}
\title{Event and root-finding functions for thresholded interventions}
\usage{
make_closure_event(response_threshold, end_threshold)
}
\arguments{
\item{response_threshold}{A named vector of length 1, with the value
representing the total number of individuals in a compartment, while the name
indicates the compartment name. The response chosen in \code{response_strategy} is
triggered when the response threshold is reached.}

\item{end_threshold}{A named vector of length 1, with the value
representing the total number of individuals in a compartment, while the name
indicates the compartment name. The response chosen in \code{response_strategy} is
ended when the end threshold is reached.}
}
\description{
Prepare functions that can be passed to \code{\link[deSolve:lsoda]{deSolve::lsoda()}} as
event and root-finding functions, which trigger and end interventions that
scale disease transmission.
}
\keyword{internal}
