% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_npi.R
\name{make_community_scaling}
\alias{make_community_scaling}
\title{Prepare community contacts scaling from workplace scaling}
\usage{
make_community_scaling(x)
}
\arguments{
\item{x}{A numeric vector of workplace contacts scaling, with each element
in the range \eqn{[0, 1]}. Expected to be length 45 (N_ECON_SECTORS), but
this is not checked.}
}
\value{
A numeric vector of length 49 (N_AGE_GROUPS + N_ECON_SECTORS).
}
\description{
Helper function to determine the (uniform) community contacts scaling,
based on the scaling of workplace contacts. Higher mean workplace contacts
scaling leads to higher community contacts scaling. Workplace contacts
scaling mean is not weighted by worker counts.
}
\keyword{internal}
