% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus_rtm.R
\name{make_conmat_large}
\alias{make_conmat_large}
\title{Make large contact matrix for Cpp model.}
\usage{
make_conmat_large(country)
}
\arguments{
\item{country}{A \verb{<daedalus_country>}.}
}
\value{
A 49x49 contact matrix.
}
\description{
Make large contact matrix for Cpp model.
}
\keyword{internal}
