% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_npi.R
\name{make_full_openness}
\alias{make_full_openness}
\title{Make a dummy openness matrix for dummy NPIs}
\usage{
make_full_openness(settings = 2)
}
\arguments{
\item{settings}{The number of contact settings. Defaults to 2.}
}
\value{
A matrix of 1.0s with dims 49 x \code{settings}.
}
\description{
Make a dummy openness matrix for dummy NPIs
}
\keyword{internal}
