% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helpers.R
\name{make_initial_state2}
\alias{make_initial_state2}
\title{Generate initial state for daedalus()}
\usage{
make_initial_state2(country, initial_state_manual = list(p_infectious = 1e-07))
}
\arguments{
\item{country}{A country or territory object of class \verb{<daedalus_country>},
\strong{or} a country or territory name from those included in the package;
see \link{country_names}, \strong{or} a country ISO2 or ISO3 code; see
\link{country_codes_iso2c} and \link{country_codes_iso3c} .
Country-specific data such as the community and workplace contacts, the
demography, and the distribution of the workforce into economic sectors is
automatically accessed from package data for the relevant country name if it
is passed as a string.
To override package defaults for country characteristics, pass a
\verb{<daedalus_country>} object instead. See \code{\link[=daedalus_country]{daedalus_country()}} for more.}

\item{initial_state_manual}{An optional \strong{named} list with the names
\code{p_infectious} and \code{p_asymptomatic} for the proportion of infectious and
symptomatic individuals in each age group and economic sector.
Defaults to \code{1e-6} and \code{0.0} respectively.}
}
\description{
Generate initial state for daedalus()
}
\keyword{internal}
