% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closure.R
\name{make_response_threshold_event}
\alias{make_response_threshold_event}
\title{Event and root-finding functions to trigger response strategies}
\usage{
make_response_threshold_event(response_threshold)
}
\arguments{
\item{response_threshold}{A single numeric value for the total number of
hospitalisations that causes an epidemic response
(specified by \code{response_strategy}) to be triggered, if it has not already
been triggered via \code{response_time}. Currently defaults to \code{NULL}, and a
\code{country}-specific spare hospital capacity value is used from
\link{country_data}.
Pass a number to override the default country-specific threshold value.}
}
\description{
Prepare functions that can be passed to \code{\link[deSolve:lsoda]{deSolve::lsoda()}} as
event and root-finding functions, which trigger response strategies that
reduce disease transmission.
}
\keyword{internal}
