% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closure.R
\name{make_rt_end_event}
\alias{make_rt_end_event}
\title{Event and root-finding functions to terminate epidemic responses}
\usage{
make_rt_end_event()
}
\description{
Prepare functions that can be passed to \code{\link[deSolve:lsoda]{deSolve::lsoda()}} as
event and root-finding functions, which trigger and end interventions that
scale disease transmission.
}
\keyword{internal}
