% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_npi.R
\name{make_workplace_scaling}
\alias{make_workplace_scaling}
\title{Prepare workplace setting scaling}
\usage{
make_workplace_scaling(x)
}
\arguments{
\item{x}{A numeric vector of workplace contacts scaling, with each element
in the range \eqn{[0, 1]}. Expected to be length 45 (N_ECON_SECTORS), but
this is not checked.}
}
\value{
A numeric vector of length 49 (N_AGE_GROUPS + N_ECON_SECTORS); this
should be c(1, 1, 1, 1, x).
}
\description{
Helper function to append a no-scaling vector of 1.0s to the front of a
workplace contacts scaling vector.
}
\keyword{internal}
