% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{model_constants}
\alias{model_constants}
\alias{N_AGE_GROUPS}
\alias{i_AGE_GROUPS}
\alias{N_VACCINE_STRATA}
\alias{N_VACCINE_DATA_GROUPS}
\alias{AGE_GROUPS}
\alias{i_WORKING_AGE}
\alias{i_SCHOOL_AGE}
\alias{N_ECON_SECTORS}
\alias{i_ECON_SECTORS}
\alias{i_EDUCATION_SECTOR}
\alias{N_ECON_STRATA}
\alias{i_NOT_WORKING}
\alias{DIM_AGE_GROUPS}
\alias{DIM_EPI_COMPARTMENTS}
\alias{DIM_ECON_SECTORS}
\alias{DIM_VACCINE_STRATA}
\alias{i_UNVACCINATED_STRATUM}
\alias{i_VACCINATED_STRATUM}
\alias{i_NEW_VAX_STRATUM}
\alias{VACCINE_GROUPS}
\alias{N_OUTPUT_COLS}
\alias{N_FLAGS}
\alias{FLAG_NAMES}
\alias{MIN_DATA_COL_NAMES}
\title{DAEDALUS model constants}
\usage{
N_AGE_GROUPS

i_AGE_GROUPS

N_VACCINE_STRATA

N_VACCINE_DATA_GROUPS

AGE_GROUPS

i_WORKING_AGE

i_SCHOOL_AGE

N_ECON_SECTORS

i_ECON_SECTORS

i_EDUCATION_SECTOR

N_ECON_STRATA

i_NOT_WORKING

DIM_AGE_GROUPS

DIM_EPI_COMPARTMENTS

DIM_ECON_SECTORS

DIM_VACCINE_STRATA

i_UNVACCINATED_STRATUM

i_VACCINATED_STRATUM

i_NEW_VAX_STRATUM

VACCINE_GROUPS

N_OUTPUT_COLS

N_FLAGS

FLAG_NAMES

MIN_DATA_COL_NAMES
}
\value{
Values for the model constants.
}
\description{
Frequently used values for the DAEDALUS model related
to the model population and structure. See also \link{epi_constants} for constants
specific to the epidemiological model.
}
\details{
The DAEDALUS model requires the following values to be fixed.
\itemize{
\item Number of age groups: 4
\item Number of vaccination strata: 2
\item Age group bins: 0-4 years, 5-19 years, 20-65 years (working age), 65+ years
\item Index of the working-age age groups: 3; see \code{AGE_GROUPS}
\item Number of economic sectors: 45
\item Index of individuals not working as a layer in the 3D state tensor: 1
\item Number of economic strata: 46; note that this is always one more than the
number of economic sectors, with the additional stratum for those not in work
which includes all individuals not of working age, and a proportion of
working age individuals.
\item Array dimension of age groups: 1
\item Array dimension of epidemiological compartments: 2
\item Array dimension of economic sectors: 3
\item Array dimension of vaccination strata: 4
\item Indices and numbers of key groups.
\item Number of state variables that are flags, switches, or indicators.
}
}
\keyword{model_constant}
