% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_hosp_overflow.R
\name{new_daedalus_hosp_overflow}
\alias{new_daedalus_hosp_overflow}
\title{Make a hospital overflow response}
\usage{
new_daedalus_hosp_overflow(country)
}
\arguments{
\item{country}{A \verb{<daedalus_country>}.}
}
\value{
An object of class \verb{<daedalus_hosp_overflow>} that inherits from
\verb{<daedalus_response>}.
}
\description{
Intended to be a helper function for internal use only. No parameters are
carried although this may change in future.
}
\keyword{internal}
