% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_output.R
\name{out_list_to_df}
\alias{out_list_to_df}
\title{Reshape output data}
\usage{
out_list_to_df(output, new_vaccinations, timesteps, labels)
}
\arguments{
\item{output}{A list-like object originating from \code{\link[=daedalus_internal]{daedalus_internal()}} and
partially processed within \code{\link[=prepare_output]{prepare_output()}}.}

\item{new_vaccinations}{A list object of new vaccination data origination in
\code{\link[=daedalus_internal]{daedalus_internal()}}.}

\item{timesteps}{A vector of timesteps, typically starting with 0.}

\item{labels}{A list of labels to be applied to the data.}
}
\value{
A \verb{<data.frame>} suitable for passing to a \verb{<daedalus_output>}
object.
}
\description{
An internal function to help \code{\link[=prepare_output]{prepare_output()}} process lists of data.
}
\keyword{internal}
