% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helpers.R
\name{prepare_contacts}
\alias{prepare_contacts}
\alias{make_conmat_large}
\alias{make_work_contacts}
\alias{make_consumer_contacts}
\alias{make_full_contacts}
\title{Make large contact matrix for Cpp model.}
\usage{
make_conmat_large(country, scaling = c("demography", "none"))

make_work_contacts(country)

make_consumer_contacts(country)

make_full_contacts(country)
}
\arguments{
\item{country}{A \verb{<daedalus_country>}.}

\item{scaling}{A string, either \code{"demography"} (default) or \code{"none"}, for
whether the matrix should be scaled by the relevant demographic (age or
economic sector) group. Scaling by demography is used as a substitute for
division by the relevant population size \eqn{N} in the force of infection
calculation in the ODE RHS, to save computation.}
}
\value{
\enumerate{
\item \code{make_conmat_large()} returns a 49x49 contact matrix scaled by the size of
demography and economic sector groups, while splitting up community contacts
from working age individuals to other age groups, among the different in
proportion to workers in those sectors.
\item \code{make_work_contacts()} returns a 45-element vector (for the number of
economic sectors) scaled by the number of workers per sector.
\item \code{make_consumer_contacts()} returns a 45x4 contact matrix with each row
scaled by the number of workers per sector. Dimensions are the number of
economic sectors and the number of age groups.
}
}
\description{
Make large contact matrix for Cpp model.
}
\keyword{internal}
