% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prepare_mutable_parameters}
\alias{prepare_mutable_parameters}
\title{Prepare mutable parameters for the DAEDALUS model}
\usage{
prepare_mutable_parameters()
}
\value{
An environment with three mutable parameters:
\itemize{
\item \code{switch}: The switch parameter which controls whether closures are active
or not.
\item \code{closures_time_start} and \code{closures_time_end}: The times at which closures
start and end. Defaults to the end time of the simulation so as to
give a default duration of 0.0.
}
}
\description{
Prepares closure start and end times for model output.
}
\keyword{internal}
