% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_country.R
\name{prepare_parameters2.daedalus_country}
\alias{prepare_parameters2.daedalus_country}
\title{Replacement prepare_parameters()}
\usage{
\method{prepare_parameters2}{daedalus_country}(x, ...)
}
\description{
Replacement prepare_parameters()
}
\keyword{internal}
