% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_country.R
\name{print.country}
\alias{print.country}
\title{Print a \verb{<country>} object}
\usage{
\method{print}{country}(x, ...)
}
\arguments{
\item{x}{An object of the \verb{<country>} class.}

\item{...}{Other parameters passed to \code{\link[=print]{print()}}.}
}
\value{
Invisibly returns the \verb{<country>} object \code{x}.
Called for printing side-effects.
}
\description{
Print a \verb{<country>} object
}
