% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reff_calculation.R
\name{reff_calculation}
\alias{reff_calculation}
\alias{get_ngm}
\title{Calculate the effective R}
\usage{
get_ngm(country, infection, p_susc = NULL)
}
\arguments{
\item{country}{A \verb{<daedalus_country>} object.}

\item{infection}{A \verb{<daedalus_infection>} object.}

\item{p_susc}{A sequence of length 4 (the number of demographic groups)
giving the proportion of susceptibles in each demographic group.}
}
\value{
A numeric matrix of dimensions \eqn{[4, 4]} where the number of rows
and columns is the number of demographic groups.
}
\description{
A simple calculation for the next-generation matrix, allowing for
heterogeneity in social contacts and optionally, a reduction in the
proportion of each age group remaining susceptible.

This assumes a single susceptible stratum with full susceptibility, and will
be updated in future versions to account for the addition of vaccination
strata with reduced susceptibility.
}
\details{
Follows the methods in Diekmann and Hesterbeek (2010) J. Roy. Soc. Interface,
<doi.org/10.1098/rsif.2009.0386>.
}
\keyword{internal}
