% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_vaccination.R
\name{scale_nu}
\alias{scale_nu}
\title{Scale vaccination rate by remaining eligibles}
\usage{
scale_nu(state, nu, uptake_limit)
}
\arguments{
\item{state}{The system state as a 4-dimensional array.}

\item{nu}{The vaccination rate.}
}
\value{
The scaled vaccination rate.
}
\description{
Scale vaccination rate by remaining eligibles
}
\keyword{internal}
