% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_gva_data.R
\docType{data}
\name{sector_gva_data}
\alias{sector_gva_data}
\title{Country economic sector value-added data}
\format{
\subsection{\code{sector_gva_data}}{

A list with 69 elements, each corresponding to a recognised country or
territory; each element is a numeric vector of length 45 (for the number of
economic sectors) which gives the daily GVA per sector for that country.
}
}
\source{
Data collected from the OECD in 2018. See processing details in
\code{data-raw/sector_gva_data.R}
}
\usage{
sector_gva_data
}
\description{
Country-wise data on gross value added (GVA) per economic sector
, taken from OECD reporting.
}
\keyword{datasets}
