% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus_multi.R
\name{split_multi_soln}
\alias{split_multi_soln}
\title{Split a dust2 solution with multiple groups}
\usage{
split_multi_soln(x)
}
\arguments{
\item{x}{A grouped dust2 output.}
}
\description{
Creates a list of solutions similar to output from a single group run.
}
\keyword{internal}
