% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_behaviour.R
\name{validate_behaviour_input}
\alias{validate_behaviour_input}
\title{Internal helper function to prepare parameters}
\usage{
validate_behaviour_input(x)
}
\description{
Internal helper function to prepare parameters
}
\keyword{internal}
