% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_country.R
\name{validate_contact_matrix}
\alias{validate_contact_matrix}
\title{Validate contact matrix}
\usage{
validate_contact_matrix(x, name)
}
\arguments{
\item{x}{An object to be validated as a contact matrix or list of matrices.}

\item{name}{A country name}
}
\value{
If \code{x} is a list of numeric matrices, or a numeric matrix
meeting expectations for Daedalus contact matrices, or \code{NULL}, returns a
list of contact matrices. If a single matrix was passed, returns a
single-element list with the name \code{"total"}.

If \code{x} is \code{NULL}, returns the country default contact matrix from
\pkg{daedalus.data} as a single-element list.

Returns errors if expectations fail.
}
\description{
Validate contact matrix
}
\keyword{internal}
