% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_country.R
\name{validate_country_input}
\alias{validate_country_input}
\title{Validate country input}
\usage{
validate_country_input(x)
}
\arguments{
\item{x}{An object to be validated as suitable as input for the
\code{country} argument of \code{\link[=daedalus]{daedalus()}}.}
}
\value{
A \verb{<daedalus_country>} object.
}
\description{
Validate country input
}
\keyword{internal}
