% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_infection.R
\name{validate_infection_input}
\alias{validate_infection_input}
\alias{validate_infection_list_input}
\title{Validate and return infection input}
\usage{
validate_infection_input(x)

validate_infection_list_input(x)
}
\arguments{
\item{x}{An object to be validated as input to the \code{infection} argument
of \code{\link[=daedalus]{daedalus()}}.}
}
\value{
\code{validate_infection_input()} returns a \verb{<daedalus_infection>}.
\code{validate_infection_list_input()} returns a list of \verb{<daedalus_infection>}s.
}
\description{
Validate and return infection input
}
\keyword{internal}
