% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_npi.R
\name{validate_openness}
\alias{validate_openness}
\title{Check openness matrix}
\usage{
validate_openness(x, settings)
}
\arguments{
\item{x}{A numeric vector or matrix to be validated as an accepted way of
specifying contacts scaling.}

\item{settings}{A number for the number of settings}
}
\value{
A numeric matrix of dimensions 49 x settings. Throws errors if \code{x}
does not conform to expectations.
}
\description{
Check openness matrix
}
\keyword{internal}
