% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helpers.R
\name{values_to_state}
\alias{values_to_state}
\title{Reshape a vector to the dimensions of the DAEDALUS state array}
\usage{
values_to_state(x)
}
\arguments{
\item{x}{A vector of numeric values.}
}
\value{
An array of dimensions (4, 9, 46, 3).
}
\description{
Reshape a vector to the dimensions of the DAEDALUS state array
}
\keyword{internal}
