% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_country.R
\name{class_country}
\alias{class_country}
\alias{country}
\alias{is_country}
\alias{print.country}
\alias{[.country}
\alias{[[.country}
\alias{[[<-.country}
\alias{$<-.country}
\title{Represent countries and territories for DAEDALUS}
\usage{
country(
  name,
  parameters = list(contact_matrix = NULL, contacts_workplace = NULL,
    contacts_consumer_worker = NULL)
)

is_country(x)

\method{print}{country}(x, ...)

\method{[}{country}(x, i)

\method{[[}{country}(x, i)

\method{[[}{country}(x, i) <- value

\method{$}{country}(x, i) <- value
}
\arguments{
\item{name}{A string giving the country or territory name; must be from among
\link{country_names}.}

\item{parameters}{An optional named list of country parameters that are
allowed to be modified. Currently, users may only pass their own contact
matrix, workplace contacts, and consumer-worker contact matrix. If these are
not passed, default values are accessed from stored package data.}

\item{x}{An object of the \verb{<country>} class.}

\item{...}{Other parameters passed to \code{\link[=print]{print()}}.}
}
\value{
\itemize{
\item \code{country()} returns an object of the S3 class \verb{<country>}.
\item \code{is_country()} returns a logical for whether an object is a \verb{<country>}.
\item Access operators \code{[} and \code{[[} return class members of a \verb{<country>} with
behaviour and return type identical to methods for a \verb{<list>}.
\item Assignment operators \verb{[[<-} and \verb{$<-} modify class members of a \verb{<country>}
and will demote the \verb{<country>} to a list if an assignment invalidates the
class.
\item \code{print.country()} invisibly returns the \verb{<country>} object \code{x}. Called for
printing side-effects.
}
}
\description{
Helper functions to create and work with S3 class \verb{<country>}
objects for use with \code{\link[=daedalus]{daedalus()}}.
These objects store country parameters for reuse and have methods for easy
parameter access and editing, as well as processing raw country
characteristics for the DAEDALUS model.
}
\examples{
x <- country("Canada")

x

country(
  "United Kingdom",
  parameters = list(contact_matrix = matrix(1, 4, 4))
)

# check whether `x` is a <country> object
is_country(x)

# assign class members
# using set_data()
set_data(x, contact_matrix = matrix(99, 4, 4))

# using assignment operators
x$contact_matrix <- matrix(99, 4, 4)
x
}
