% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_output.R
\name{class_daedalus_output}
\alias{class_daedalus_output}
\alias{as_daedalus_output}
\alias{validate_daedalus_output}
\alias{is_daedalus_output}
\alias{print.daedalus_output}
\title{Convert a list to \verb{<daedalus_output>}}
\usage{
as_daedalus_output(x)

validate_daedalus_output(x)

is_daedalus_output(x)

\method{print}{daedalus_output}(x, ...)
}
\arguments{
\item{x}{An object of the \verb{<daedalus_output>} class.}
}
\value{
An object of the \verb{<daedalus_output>} class.

Invisibly returns \code{x}; called primarily for its side effects of
erroring when the object does not satisfy the \verb{<daedalus_output>} class
requirements.

A logical for whether \code{x} is of the \verb{<daedalus_output>} class.

None; called for its printing side effects.
}
\description{
Convert a list to \verb{<daedalus_output>}

Validate a potential \verb{<daedalus_output>} class object

Check if an object is of the \verb{<daedalus_output>} class

Print \verb{<daedalus_output>} class objects
}
\keyword{internal}
