% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_infection.R
\name{class_infection}
\alias{class_infection}
\alias{infection}
\alias{is_infection}
\alias{print.infection}
\alias{[.infection}
\alias{[[.infection}
\alias{[[<-.infection}
\alias{$<-.infection}
\title{Represent countries and territories for DAEDALUS}
\usage{
infection(name, ...)

is_infection(x)

\method{print}{infection}(x, ...)

\method{[}{infection}(x, i)

\method{[[}{infection}(x, i)

\method{[[}{infection}(x, i) <- value

\method{$}{infection}(x, i) <- value
}
\arguments{
\item{name}{An epidemic name from among \link{epidemic_names}.
Selecting an epidemic automatically pulls in infection parameters
associated with the epidemic; these are stored as packaged data in
\code{daedalus::infection_data}. Default infection parameters for epidemics can be
over-ridden by passing them as a named list to \code{...}.}

\item{...}{Other parameters passed to \code{\link[=print]{print()}}.}

\item{x}{An object of the \verb{<infection>} class.}

\item{i}{The index or name to access.}

\item{value}{The value to assign to index or name \code{i}.}
}
\value{
\itemize{
\item \code{infection()} returns an object of the S3 class \verb{<infection>}.
\item \code{is_infection()} returns a logical for whether an object is a
\verb{<infection>}.
\item Access operators \code{[} and \code{[[} return class members of a \verb{<infection>} with
behaviour and return type identical to methods for a \verb{<list>}.
\item Assignment operators \verb{[[<-} and \verb{$<-} modify class members of a
\verb{<infection>} and will demote the \verb{<infection>} to a list if an assignment
invalidates the class.
\item \code{print.infection()} invisibly returns the \verb{<infection>} object \code{x}.
Called for printing side-effects.
}
}
\description{
Helper functions to create and work with S3 class \verb{<infection>}
objects for use with \code{\link[=daedalus]{daedalus()}}.
These objects store infection parameters for reuse and have methods for easy
parameter access and editing, as well as processing raw infection
characteristics for the DAEDALUS model.
}
\details{
\subsection{Included epidemics}{

Epidemics for which data are available are given below (pathogen in
parentheses). The string indicates the name that must be passed to the \code{name}
argument.
\itemize{
\item \code{"sars_cov_1"}: SARS 2004 (SARS-CoV-1),
\item \code{"influenza_2009"}: influenza 2009 (influenza A H1N1),
\item \code{"influenza_1957"}: influenza 1957 (influenza A H2N2),
\item \code{"influenza_1918"}: influenza 1918 (influenza A H1N1),
\item \code{"sars_cov_2_pre_alpha"}: Covid-19 wild type (SARS-Cov-2 wild type),
\item \code{"sars_cov_2_omicron"}: Covid-19 Omicron (SARS-CoV-2 omicron),
\item \code{"sars_cov_2_pre_delta"}: (SARS-CoV-2 delta).
}
}

\subsection{Infection parameters}{

All infections have the following parameters, which take default values
stored in the package under \link{infection_data}. Users can pass
custom values for these parameters as arguments via \code{...}.
\itemize{
\item \code{r0}: A single numeric value for the basic reproduction value of the
infection \eqn{R_0}.
\item \code{sigma}: A single numeric value > 0.0 for the rate of transition from the
exposed compartment to one of two infectious compartments.
\item \code{p_sigma}: A single numeric value in the range \eqn{[0.0, 1.0]} for the
proportion of infectious individuals who are also symptomatic. Asymptomatic
individuals can have a different contribution to the force of infection from
symptomatic individuals.
\item \code{epsilon}: A single numeric value for the relative contribution of
asymptomatic infectious individuals to the force of infection (compared to
symptomatic individuals).
\item \code{gamma_Is}: A single numeric value for the recovery rate of infectious
individuals who are not hospitalised.
\item \code{gamma_Ia}: A single numeric value for the recovery rate from asymptomatic
infection.
\item \code{gamma_H}: A numeric vector of length \code{N_AGE_GROUPS} (4) for the
age-specific recovery rate for individuals who are hospitalised.
\item \code{eta}: A numeric vector of length \code{N_AGE_GROUPS} (4) for the age-specific
hospitalisation rate for individuals who are infectious and symptomatic.
\item \code{omega}: A numeric vector of length \code{N_AGE_GROUPS} (4) for the age-specific
mortality rate for individuals who are hospitalised.
\item \code{rho}: A single numeric value for the rate at which infection-derived
immunity wanes, returning individuals in the 'recovered' compartment to the
'susceptible' compartment.
}
}
}
\examples{
# make an <infection> object with default parameter values
infection("sars_cov_1")

# modify infection parameters R0 and immunity waning rate
infection("influenza_1918", r0 = 2.5, rho = 0.01)
}
