% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_npi.R, R/class_timed_npi.R
\name{class_npi}
\alias{class_npi}
\alias{daedalus_npi}
\alias{is_daedalus_npi}
\alias{print.daedalus_npi}
\alias{daedalus_timed_npi}
\title{Represent non-pharmaceutical intervention strategies for DAEDALUS}
\usage{
daedalus_npi(
  name,
  country,
  infection,
  openness = NULL,
  start_time = 30,
  end_time = NULL,
  max_duration = 365
)

is_daedalus_npi(x)

\method{print}{daedalus_npi}(x, ...)

daedalus_timed_npi(start_time, end_time, openness, country)
}
\arguments{
\item{name}{An NPI strategy name from among
\link[daedalus.data:closure_strategies]{daedalus.data::closure_strategy_names}, or \code{NA}.
Passing a pre-defined strategy name automatically pulls in openness
parameters for the associated response strategy; these are stored as
packaged data in \link[daedalus.data:closure_strategies]{daedalus.data::closure_strategy_data}.

Pass \code{NA} to define a custom response strategy by passing a vector to
\code{openness}.}

\item{country}{A country or territory object of class \verb{<daedalus_country>},
\strong{or} a country or territory name from those included in the package;
see \link[daedalus.data:country_names_codes]{daedalus.data::country_names}, \strong{or} a country ISO2 or ISO3 code; see
\link[daedalus.data:country_names_codes]{daedalus.data::country_codes_iso2c} and
\link[daedalus.data:country_names_codes]{daedalus.data::country_codes_iso3c}.
Country-specific data such as the community and workplace contacts, the
demography, and the distribution of the workforce into economic sectors is
automatically accessed from package data for the relevant country name if it
is passed as a string.
To override package defaults for country characteristics, pass a
\verb{<daedalus_country>} object instead. See \code{\link[=daedalus_country]{daedalus_country()}} for more.}

\item{infection}{An infection parameter object of the class
\verb{<daedalus_infection>}, \strong{or} an epidemic name for which data are provided
in the package; see \link[daedalus.data:epidemic_data]{daedalus.data::epidemic_names} for historical epidemics
or epidemic waves for which parameters are available.
Passing the name as a string automatically accesses the default parameters
of an infection. Create and pass a \verb{<daedalus_infection>} to tweak infection
parameters using \code{\link[=daedalus_infection]{daedalus_infection()}}.}

\item{openness}{For \code{daedalus_npi()}, an optional numeric matrix giving the openness of each
setting in the model when the NPI is in effect.

For \code{daedalus_timed_npi()}, a list of numeric matrices giving the openness
coefficients for each interval specified by corresponding elements of
\code{start_time} and \code{end_time}.

Expected to have a minimum size of \code{N_ECON_SECTORS + N_AGE_GROUPS}
(currently 98), corresponding to two settings with 49 strata per setting,
with all values are in the range \eqn{[0, 1]},}

\item{start_time}{For \code{daedalus_npi()}, a single number giving the number of days after the
start of the model that the NPI response begins. Defaults to 30.

For \code{daedalus_timed_npi()}: may be a vector of NPI start times, with no
default values.}

\item{end_time}{For \code{daedalus_npi()}, a single number giving the number of days after the
start of the model that the NPI response ends. Defaults to \code{NULL},
indicating that the response ends (if no other condition is met), at the
maximum duration.

For \code{daedalus_timed_npi()}: may be a vector of NPI end times, with no
default values. Timed NPIs have no default duration.}

\item{max_duration}{The maximum number of days that an NPI response is active
whether started by passing the \code{start_time} or when a state threshold is
crossed. Defaults to 365 days. NPIs created using \code{daedalus_timed_npi()} have
no default maximum duration and are encoded by pairs of start and end times.}

\item{x}{An object to be tested or printed as a \verb{<daedalus_npi>}.}

\item{...}{For \code{daedalus_npi()}, other parameters passed to
\code{\link[=new_daedalus_response]{new_daedalus_response()}}.
For the \code{print} method, other parameters passed to \code{\link[=print]{print()}}.}
}
\value{
\code{daedalus_npi()}: A \verb{<daedalus_npi>} class object that specifies an NPI that
may be responsive to both model state and time.

\code{daedalus_timed_npi()}: A \verb{<daedalus_npi>} class object which specifies
time-limited interventions only.
}
\description{
Helper functions to create and work with S3 class
\verb{<daedalus_npi>} objects for use with \code{\link[=daedalus]{daedalus()}}.
These objects store NPI parameters for reuse and have methods for
easy parameter access and editing, as well as processing raw NPI
characteristics for the DAEDALUS model.

Most NPIs must be initialised with an associated \code{country} and \code{infection};
these are used to determine flag positions as well as the state-values
(hospital capacity) and incidence-prevalence ratio (IPR) at which the NPI
starts and ends reactively.

\code{daedalus_timed_npi()} is a helper function to create a \verb{<daedalus_npi>} that
is only trigged by time, and is not responsive to state variables.
Primarily intended for use in real time modelling.
}
\details{
Note that NPIs created using \code{daedalus_npi()} are reactive to the model state
(i.e., the epidemic state), and can trigger multiple times when state
conditions are met.
}
\examples{
# for a school closure strategy
daedalus_npi("school_closures", "GBR", "sars_cov_1")

# set custom openness
daedalus_npi(
  NA,
  "GBR", "sars_cov_1",
  openness = cbind(rep(1, 49), rep(0.1, 49))
)

# time-limited NPI with multiple phases
phase_1 <- cbind(
  rep(1, 49),
  rep(0.5, 49)
)
phase_2 <- cbind(
  rep(1, 49),
  rep(0.3, 49)
)
phase_3 <- cbind(
  rep(1, 49),
  rep(0.8, 49)
)

daedalus_timed_npi(
  start_time = c(10, 20, 30),
  end_time = c(15, 25, 40),
  openness = list(
    phase_1, phase_2, phase_3
  ),
  country = "GBR"
)

}
