% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_npi.R
\name{class_npi}
\alias{class_npi}
\alias{daedalus_npi}
\alias{is_daedalus_npi}
\alias{print.daedalus_npi}
\title{Represent non-pharmaceutical intervention strategies for DAEDALUS}
\usage{
daedalus_npi(
  name,
  country,
  infection,
  openness = NULL,
  start_time = 30,
  end_time = NA_real_,
  max_duration = 365
)

is_daedalus_npi(x)

\method{print}{daedalus_npi}(x, ...)
}
\arguments{
\item{name}{An NPI strategy name from among
\link[daedalus.data:closure_strategies]{daedalus.data::closure_strategy_names}, or \code{NA}.
Passing a pre-defined strategy name automatically pulls in openness
parameters for the associated response strategy; these are stored as
packaged data in \code{daedalus.data::closure_strategy_data}.
Pass \code{NA} to define a custom response strategy by passing a vector to
\code{openness}.}

\item{country}{A country or territory object of class \verb{<daedalus_country>},
\strong{or} a country or territory name from those included in the package;
see \link[daedalus.data:country_names_codes]{daedalus.data::country_names}, \strong{or} a country ISO2 or ISO3 code; see
\link[daedalus.data:country_names_codes]{daedalus.data::country_codes_iso2c} and
\link[daedalus.data:country_names_codes]{daedalus.data::country_codes_iso3c}.
Country-specific data such as the community and workplace contacts, the
demography, and the distribution of the workforce into economic sectors is
automatically accessed from package data for the relevant country name if it
is passed as a string.
To override package defaults for country characteristics, pass a
\verb{<daedalus_country>} object instead. See \code{\link[=daedalus_country]{daedalus_country()}} for more.}

\item{infection}{An infection parameter object of the class
\verb{<daedalus_infection>}, \strong{or} an epidemic name for which data are provided
in the package; see \link[daedalus.data:epidemic_data]{daedalus.data::epidemic_names} for historical epidemics
or epidemic waves for which parameters are available.
Passing the name as a string automatically accesses the default parameters
of an infection. Create and pass a \verb{<daedalus_infection>} to tweak infection
parameters using \code{\link[=daedalus_infection]{daedalus_infection()}}.}

\item{openness}{An optional numeric vector where all values are in the range
\eqn{[0, 1]}, giving the openness of each economic sector in the model.
Expected to have a length of \code{N_ECON_SECTORS} (currently 45).}

\item{start_time}{The number of days after the start of the epidemic that
the NPI response begins. May be a vector. Defaults to 30. Passed to
the \code{time_on} argument in \code{\link[=new_daedalus_response]{new_daedalus_response()}} via the class constructor
\code{new_daedalus_npi()}.}

\item{end_time}{The number of days after the start of the epidemic that
the NPI response ends. May be a vector, with times taken with reference to
\code{start_time}. Defaults to \code{NA_real_}, with no defined end time. Passed to
the \code{time_off} argument in \code{\link[=new_daedalus_response]{new_daedalus_response()}} via
\code{new_daedalus_npi()}.}

\item{max_duration}{The maximum number of days that an NPI response is active
whether started by passing the \code{start_time} or when a state threshold is
crossed. Defaults to 365 days. \code{max_duration} overrides \code{end_time} in ending
NPIs, if the maximum duration is reached before \code{end_time}.}

\item{x}{An object to be tested or printed as a \verb{<daedalus_npi>}.}

\item{...}{For \code{daedalus_npi()}, other parameters passed to
\code{\link[=new_daedalus_response]{new_daedalus_response()}}.
For the \code{print} method, other parameters passed to \code{\link[=print]{print()}}.}
}
\description{
Helper functions to create and work with S3 class
\verb{<daedalus_npi>} objects for use with \code{\link[=daedalus]{daedalus()}}.
These objects store NPI parameters for reuse and have methods for
easy parameter access and editing, as well as processing raw NPI
characteristics for the DAEDALUS model.

All NPIs must be initialised with an associated \code{country} and \code{infection};
these are used to determine the state-values (hospital capacity) and
incidence-prevalence ratio (IPR) at which the NPI starts and ends reactively.
}
\details{
Note that NPIs are reactive to the model state (i.e., the epidemic state),
and can trigger when state conditions are met.
}
\examples{
# for a school closure strategy
daedalus_npi("school_closures", "GBR", "sars_cov_1")

# set custom openness
daedalus_npi(
  NA,
  "GBR", "sars_cov_1",
  openness = rep(0.1, 45)
)
}
