% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_response.R
\name{class_response}
\alias{class_response}
\alias{new_daedalus_response}
\alias{new_daedalus_npi}
\alias{new_daedalus_behaviour}
\alias{new_daedalus_mortality}
\alias{is_daedalus_response}
\alias{validate_daedalus_response}
\title{Constructor for the <daedalus_response> super-class and sub-classes}
\usage{
new_daedalus_response(
  name,
  class,
  parameters,
  time_on = NULL,
  duration = NULL,
  id_state_on = NULL,
  value_state_on = NULL,
  id_state_off = NULL,
  value_state_off = NULL
)

new_daedalus_npi()

new_daedalus_behaviour()

new_daedalus_mortality()

is_daedalus_response(x)

validate_daedalus_response(x)
}
\arguments{
\item{name}{Response name. This will eventually be passed to the name
argument in the C++ class \verb{daedalus::events::response}. No default for the
super-class constructor; each sub-class has a fixed name to aid
identifiability in \code{dust2} events outputs.}

\item{class}{The name of the sub-class being created. May be one of
"daedalus_npi", "daedalus_vaccination", "daedalus_behaviour", or
"daedalus_mortality".}

\item{parameters}{Sub-class parameters passed from sub-class constructors.}

\item{time_on}{Intended to be a numeric vector of start times.}

\item{duration}{Intended to be a numeric vector of durations.}

\item{id_state_on}{Intended to be a numeric vector of compartment indices
which are checked for roots that when found launch an event. The check is
always performed on the \emph{sum} of the state at the indices.}

\item{value_state_on}{Intended to be a numeric vector of roots which when
found change a switch-like state-flag to 'on'. The check is
always performed on the \emph{sum} of the state at the indices.}

\item{id_state_off}{Intended to be a numeric vector of compartment indices
which are checked for roots that when found end an event.}

\item{value_state_off}{Intended to be a numeric vector of roots which when
found change a switch-like state-flag to 'off'.}

\item{x}{For all functions taking \code{x}, must be an object of class
\verb{<daedalus_response>}, or an object to be validated as of the class.}
}
\value{
\code{new_daedalus_response()} returns a \verb{<daedalus_response>}. Since there is no
default value to \code{class}, each \verb{<daedalus_response>} must also be of one of
the sub-classes returned by the sub-class constructors:
\itemize{
\item \code{new_daedalus_npi()}: \verb{<daedalus_npi>};
\item \code{new_daedalus_vax()}: \verb{<daedalus_vax>};
\item \code{new_daedalus_behaviour()}: \verb{<daedalus_behaviour>};
\item \code{new_daedalus_mortality()}: \verb{<daedalus_mortality>}.
}
}
\description{
Constructor for the <daedalus_response> super-class and sub-classes
}
\keyword{internal}
