% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_response.R
\docType{data}
\name{class_response}
\alias{class_response}
\alias{new_daedalus_response}
\alias{response_class_names}
\alias{new_daedalus_behaviour}
\alias{is_daedalus_response}
\alias{validate_daedalus_response}
\title{Constructor for the <daedalus_response> super-class and sub-classes}
\format{
An object of class \code{character} of length 4.
}
\usage{
new_daedalus_response(
  name,
  class = response_class_names,
  parameters = list(),
  identifier = NA_character_,
  id_flag = NA_integer_,
  time_on = NA_real_,
  time_off = NA_real_,
  max_duration = NA_real_,
  id_state_on = NA_integer_,
  value_state_on = NA_real_,
  id_state_off = NA_integer_,
  value_state_off = NA_real_,
  root_state_on = 1L,
  root_state_off = -1L,
  id_time_log = NA_integer_
)

response_class_names

new_daedalus_behaviour()

is_daedalus_response(x)

validate_daedalus_response(x)
}
\arguments{
\item{name}{Response name. This will eventually be passed to the name
argument in the C++ class \verb{daedalus::events::response}. No default for the
super-class constructor; each sub-class has a fixed name to aid
identifiability in \code{dust2} events outputs.}

\item{class}{The name of the sub-class being created. May be one of
"daedalus_npi", "daedalus_vaccination", "daedalus_behaviour", or
"daedalus_mortality". No default argument, users are forced to specify a
class when building off this constructor.}

\item{parameters}{Sub-class parameters passed from sub-class constructors.
Defaults to an empty list to aid development.}

\item{identifier}{An optional string giving the name of a pre-defined
response strategy. Sub-class constructors and helper functions provide the
names and handle cases where the strategy is a custom one.}

\item{id_flag}{Index for the state-flag that should be changed when this
response is on or off. Typically auto-calculated by sub-class constructor or
helper functions; see e.g. \code{\link[=daedalus_vaccination]{daedalus_vaccination()}}.}

\item{time_on}{Intended to be a numeric vector of start times.}

\item{time_off}{Intended to be a numeric vector of end times.}

\item{max_duration}{Intended to be a number for the maximum duration.}

\item{id_state_on}{Intended to be a numeric vector of compartment indices
which are checked for roots that when found launch an event. The check is
always performed on the \emph{sum} of the state at the indices.}

\item{value_state_on}{Intended to be a numeric vector of roots which when
found change a switch-like state-flag to 'on'. The check is
always performed on the \emph{sum} of the state at the indices.}

\item{id_state_off}{Intended to be a numeric vector of compartment indices
which are checked for roots that when found end an event.}

\item{value_state_off}{Intended to be a numeric vector of roots which when
found change a switch-like state-flag to 'off'.}

\item{root_state_on}{A single integer, either 1 or -1, for whether
state-dependent launch is on an increasing or decreasing root, respectively.
Defaults to 1 for an increasing root as this is the most common use case.}

\item{root_state_off}{A single integer, either 1 or -1, for whether
state-dependent end is on an increasing or decreasing root, respectively.
Defaults to -1 for a decreasing root as this is the most common use case.}

\item{id_time_log}{Intended to be a single number for the state index where
the start-time of this response is stored. See also \code{\link[=initial_flags]{initial_flags()}}. This
attribute is intended solely to enable time-limitation on state-triggers.}

\item{x}{For all functions taking \code{x}, must be an object of class
\verb{<daedalus_response>}, or an object to be validated as of the class.}
}
\value{
\code{new_daedalus_response()} returns a \verb{<daedalus_response>}. Since there is no
default value to \code{class}, each \verb{<daedalus_response>} must also be of one of
the sub-classes returned by the sub-class constructors:
\itemize{
\item \code{new_daedalus_npi()}: \verb{<daedalus_npi>};
\item \code{new_daedalus_vax()}: \verb{<daedalus_vax>};
\item \code{new_daedalus_behaviour()}: \verb{<daedalus_behaviour>};
\item \code{new_daedalus_hosp_overflow()}: \verb{<daedalus_hosp_overflow>}. This function
is not exposed to users.
}
}
\description{
Constructor for the <daedalus_response> super-class and sub-classes
}
\keyword{datasets}
\keyword{internal}
