% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_vaccination.R
\name{class_vaccination}
\alias{class_vaccination}
\alias{daedalus_vaccination}
\alias{is_daedalus_vaccination}
\alias{print.daedalus_vaccination}
\title{Represent vaccine investment scenarios for DAEDALUS}
\usage{
daedalus_vaccination(name, ...)

is_daedalus_vaccination(x)

\method{print}{daedalus_vaccination}(x, ...)
}
\arguments{
\item{name}{A vaccination investment scenario name from among
\link{vaccination_scenario_names}.
Selecting an epidemic automatically pulls in vaccination parameters
associated with the epidemic; these are stored as packaged data in
\code{daedalus::vaccination_scenario_data}. Default vaccination parameters can be
over-ridden by passing them as a named list to \code{...}.}

\item{...}{Other parameters passed to \code{\link[=print]{print()}}.}

\item{x}{An object of the \verb{<daedalus_vaccination>} class.}
}
\description{
Helper functions to create and work with S3 class
\verb{<daedalus_vaccination>} objects for use with \code{\link[=daedalus]{daedalus()}}.
These objects store vaccination parameters for reuse and have methods for
easy parameter access and editing, as well as processing raw vaccination
characteristics for the DAEDALUS model.
}
\section{Vaccination parameters}{
\itemize{
\item \code{vax_start_time}: The number of days after the start of the epidemic that
vaccination begins. Must be a single number.
\item \code{nu}: A single number for the percentage of the total population that can
be vaccinated each day. This is converted into a proportion automatically
within \code{\link[=daedalus]{daedalus()}}.
\item \code{vax_uptake_limit}: A single number giving the upper limit for the
percentage of the population that can be vaccinated. When this limit is
reached, the vaccination rate \code{nu} is set to zero.
}
}

