% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus-package.R
\docType{package}
\name{daedalus-package}
\alias{daedalus-package}
\title{daedalus: Model health, social, and economic costs of a pandemic using _DAEDALUS_}
\description{
Model the health, social, and economic costs of directly transmitted respiratory virus pandemics, under different scenarios of pro-active preparedness and reactive interventions, using the _DAEDALUS_ integrated health-economics model adapted from Haw et al. (2022) <doi.org/10.1038/s43588-022-00233-0>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/j-idea/daedalus}
  \item Report bugs at \url{https://github.com/j-idea/daedalus/issues}
}

}
\author{
\strong{Maintainer}: Pratik Gupte \email{p.gupte24@imperial.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID})

Other contributors:
\itemize{
  \item Imperial College of Science, Technology and Medicine [copyright holder]
}

}
\keyword{internal}
