% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus-package.R
\docType{package}
\name{daedalus-package}
\alias{daedalus-package}
\title{daedalus: Optimise economic, social and health trade-offs in a pandemic}
\description{
Model and optimise the trade-offs between social and economic activity while containing a pandemic, as presented in Haw et al. (2022) <doi.org/10.1038/s43588-022-00233-0>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/j-idea/daedalus}
  \item Report bugs at \url{https://github.com/j-idea/daedalus/issues}
}

}
\author{
\strong{Maintainer}: Pratik Gupte \email{p.gupte24@imperial.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID})

Other contributors:
\itemize{
  \item Imperial College of Science, Technology and Medicine [copyright holder]
}

}
\keyword{internal}
