% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus.R
\name{daedalus}
\alias{daedalus}
\title{DAEDALUS model for health, social and economic costs of a pandemic}
\usage{
daedalus(
  country,
  infection,
  response_strategy = NULL,
  vaccine_investment = NULL,
  response_time = 30,
  response_duration = 365,
  initial_state_manual = NULL,
  time_end = 600,
  ...
)
}
\arguments{
\item{country}{A country or territory object of class \verb{<daedalus_country>},
\strong{or} a country or territory name from those included in the package;
see \link[daedalus.data:country_names_codes]{daedalus.data::country_names}, \strong{or} a country ISO2 or ISO3 code; see
\link[daedalus.data:country_names_codes]{daedalus.data::country_codes_iso2c} and
\link[daedalus.data:country_names_codes]{daedalus.data::country_codes_iso3c}.
Country-specific data such as the community and workplace contacts, the
demography, and the distribution of the workforce into economic sectors is
automatically accessed from package data for the relevant country name if it
is passed as a string.
To override package defaults for country characteristics, pass a
\verb{<daedalus_country>} object instead. See \code{\link[=daedalus_country]{daedalus_country()}} for more.}

\item{infection}{An infection parameter object of the class
\verb{<daedalus_infection>}, \strong{or} an epidemic name for which data are provided
in the package; see \link[daedalus.data:epidemic_data]{daedalus.data::epidemic_names} for historical epidemics
or epidemic waves for which parameters are available. Alternatively, a list
of \verb{<daedalus_infection>} objects.

Passing the name as a string automatically accesses the default parameters
of an infection. Create and pass a \verb{<daedalus_infection>} to tweak infection
parameters using \code{\link[=daedalus_infection]{daedalus_infection()}}.}

\item{response_strategy}{A string for the name of response strategy followed;
defaults to "none". The response strategy determines the country-specific
response threshold following which the response is activated. See
\code{response_threshold}.

While the response strategy is active, economic contacts are scaled using the
package data object \code{daedalus.data::closure_data}.}

\item{vaccine_investment}{Either a single string or a
\verb{<daedalus_vaccination>} object specifying the vaccination parameters
associated with an advance vaccine-investment scenario. Defaults to \code{NULL}
for absolutely no vaccination in the model.
A vaccination investment of \code{"none"} indicates no \emph{prior} investment, but
the model will include vaccination beginning after 1 year, at a low rate
across all age groups. Other accepted values are \code{"low"}, \code{"medium"} and
\code{"high"}. See \code{\link[=daedalus_vaccination]{daedalus_vaccination()}} for more information.}

\item{response_time}{A single numeric value for the time in days
at which the selected response is activated. This is ignored if the response
has already been activated by the hospitalisation threshold being reached.
Defaults to 30 days.}

\item{response_duration}{A single integer-ish number that gives the number of
days after \code{response_time} that an NPI should end.}

\item{initial_state_manual}{An optional \strong{named} list with the names
\code{p_infectious} and \code{p_asymptomatic} for the proportion of infectious and
symptomatic individuals in each age group and economic sector.
Defaults to \code{1e-6} and \code{0.0} respectively.}

\item{time_end}{An integer-like value for the number of timesteps
at which to return data. This is treated as the number of days with data
returned for each day. Defaults to 300 days.}

\item{...}{Optional arguments that are passed to \code{\link[dust2:dust_ode_control]{dust2::dust_ode_control()}}.}
}
\value{
A \verb{<daedalus_output>} object if \code{infection} is a string or a single
\verb{<daedalus_infection>} object. Otherwise, a list of \verb{<daedalus_output>}s
of the same length of \code{infection} if a list of \verb{<daedalus_infection>}s is
passed to \code{infection}.
}
\description{
Run the DAEDALUS model from R. This is a work in progress.
}
\details{
Users can pass the following initial state parameters to
\code{initial_state_manual}:
\itemize{
\item \code{p_infectious}: A single numeric value in the range \eqn{[0.0, 1.0]} giving
the proportion of individuals in each age group and economic sector that are
to be initialised as infectious. Defaults to \code{1e-6}, or one in every one
million as infectious.
\item \code{p_asymptomatic}: A single numeric value in the range \eqn{[0.0, 1.0]} for
the proportion of initially infectious individuals who are considered to be
asymptomatic. Defaults to 0.0.
}
}
\examples{
# country and infection specified by strings using default characteristics
output <- daedalus(
  "Canada", "influenza_1918"
)

# country passed as <daedalus_country> with some characteristics modified
country_x <- daedalus_country(
  "Canada",
  parameters = list(contact_matrix = matrix(5, 4, 4)) # uniform contacts
)
output <- daedalus(country_x, "influenza_1918")

# with some infection parameters over-ridden by the user
output <- daedalus(
  "United Kingdom",
  daedalus_infection("influenza_1918", r0 = 1.3)
)

# with default initial conditions over-ridden by the user
output <- daedalus(
  "United Kingdom", "influenza_1918",
  initial_state_manual = list(p_infectious = 1e-3)
)

}
