% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus.R
\name{daedalus}
\alias{daedalus}
\title{DAEDALUS model for health, social and economic costs of a pandemic}
\usage{
daedalus(
  country,
  infection,
  response_strategy = NULL,
  vaccine_investment = NULL,
  response_time = 30,
  auto_social_distancing = c("off", "independent", "npi_linked"),
  initial_state_manual = NULL,
  time_end = 600,
  ...
)
}
\arguments{
\item{country}{A country or territory object of class \verb{<daedalus_country>},
\strong{or} a country or territory name from those included in the package;
see \link[daedalus.data:country_names_codes]{daedalus.data::country_names}, \strong{or} a country ISO2 or ISO3 code; see
\link[daedalus.data:country_names_codes]{daedalus.data::country_codes_iso2c} and
\link[daedalus.data:country_names_codes]{daedalus.data::country_codes_iso3c}.
Country-specific data such as the community and workplace contacts, the
demography, and the distribution of the workforce into economic sectors is
automatically accessed from package data for the relevant country name if it
is passed as a string.
To override package defaults for country characteristics, pass a
\verb{<daedalus_country>} object instead. See \code{\link[=daedalus_country]{daedalus_country()}} for more.}

\item{infection}{An infection parameter object of the class
\verb{<daedalus_infection>}, \strong{or} an epidemic name for which data are provided
in the package; see \link[daedalus.data:epidemic_data]{daedalus.data::epidemic_names} for historical epidemics
or epidemic waves for which parameters are available.
Passing the name as a string automatically accesses the default parameters
of an infection. Create and pass a \verb{<daedalus_infection>} to tweak infection
parameters using \code{\link[=daedalus_infection]{daedalus_infection()}}.}

\item{response_strategy}{A string for the name of response strategy followed,
a numeric of length 45 (number of economic sectors), or a \verb{<daedalus_npi>}
object. Defaults to "none".
While the response strategy is active, economic contacts are scaled using the
package data object \code{daedalus.data::closure_strategy_data}.}

\item{vaccine_investment}{Either a single string or a
\verb{<daedalus_vaccination>} object specifying the vaccination parameters
associated with an advance vaccine-investment scenario. Defaults to \code{NULL}
for absolutely no vaccination in the model.
A vaccination investment of \code{"none"} indicates no \emph{prior} investment, but
the model will include vaccination beginning after 1 year, at a low rate
across all age groups. Other accepted values are \code{"low"}, \code{"medium"} and
\code{"high"}. See \code{\link[=daedalus_vaccination]{daedalus_vaccination()}} for more information.}

\item{response_time}{A single numeric value for the time in days
at which the selected response is activated. This is ignored if the response
has already been activated by the hospitalisation threshold being reached.
Defaults to 30 days. Responses have a default maximum duration of 365 days.
This can be changed by passing a \verb{<daedalus_npi>} object to
\code{response_strategy}.}

\item{auto_social_distancing}{A string giving the option for the form of
spontaneous social distancing in the model, which reduces infection
transmission as a function of daily deaths. See \strong{Details} for more.}

\item{initial_state_manual}{An optional \strong{named} list with the names
\code{p_infectious}, \code{p_asymptomatic}, and \code{p_immune}.
\code{p_infectious} and \code{p_asymptomatic} give the proportion of
infectious and symptomatic individuals in each age group and economic sector.
Defaults to \code{1e-6} and \code{0.0} respectively.
\code{p_immune} may be a single number in the range \verb{0.0 <= p_immune <= 1.0} or a
4-element vector in that range (the number of age groups in the model), for
the proportion of individuals in the population or in each age group that
have some pre-existing immunity to infection (reduced susceptibility). See
\strong{Details} for more.}

\item{time_end}{An integer-like value for the number of timesteps
at which to return data. This is treated as the number of days with data
returned for each day. Defaults to 300 days.}

\item{...}{Optional arguments that are passed to \code{\link[dust2:dust_ode_control]{dust2::dust_ode_control()}}.}
}
\value{
A \verb{<daedalus_output>} object if \code{infection} is a string or a single
\verb{<daedalus_infection>} object. Otherwise, a list of \verb{<daedalus_output>}s
of the same length of \code{infection} if a list of \verb{<daedalus_infection>}s is
passed to \code{infection}.
}
\description{
Run the DAEDALUS model from R. This is a work in progress.
}
\details{
\subsection{Details: Initial state}{

Users can pass the following initial state parameters to
\code{initial_state_manual}:
\itemize{
\item \code{p_infectious}: A single numeric value in the range \eqn{[0.0, 1.0]} giving
the proportion of individuals in each age group and economic sector that are
to be initialised as infectious. Defaults to \code{1e-6}, or one in every one
million as infectious.
\item \code{p_asymptomatic}: A single numeric value in the range \eqn{[0.0, 1.0]} for
the proportion of initially infectious individuals who are considered to be
asymptomatic. Defaults to 0.0.
\item \code{p_immune}: Either a single number or a vector of 4 elements (the number
of age groups) in the range \eqn{[0.0, 1.0]} for the proportion of the
population (or each age group) that is considered to have pre-existing
immunity. This is a stop-gap implementation that assumes one of two cases:
(1) if no vaccination is intended in the model and \code{vaccine_investment} is
\code{NULL}, the susceptibility of individuals pre-existing immunity is 50\%; or
(2) if a vaccination strategy is specified, the pre-existing immunity is
assumed to be from a prior rollout, and the susceptibility is determined by
the chosen vaccination strategy (as \code{1 - efficacy}).
}
}

\subsection{Details: Spontaneous social distancing}{

There are three possible options for this behavioural module, given below.
\strong{Note that} a major issue with including this in a model run (any value
other than \code{"off"}) is that it leads to substantially lower response costs,
and generally better health outcomes (lives lost), \strong{without accounting for
any attendant economic or social costs}. As such, please treat this option
as \strong{highly experimental}.
\itemize{
\item \code{"off"}: There is no effect of daily deaths on infection transmissibility.
This is the \strong{default choice}.
\item \code{"independent"}: Public-concern over deaths reduces transmissibility, and
is independent of any mandated responses. It begins at the start of the
simulation, and continues until the simulation ends.
\item \verb{"npi_linked}": Public-concern over deaths reduces transmissibility, but
only when a  mandated response is active. Note that there is currently no way
to end a response triggered by a state, i.e., an NPI launched due to hospital
capacity being breached.
}
}
}
\examples{
# country and infection specified by strings using default characteristics
output <- daedalus(
  "Canada", "influenza_1918"
)

# country passed as <daedalus_country> with some characteristics modified
country_x <- daedalus_country(
  "Canada",
  parameters = list(contact_matrix = matrix(5, 4, 4)) # uniform contacts
)
output <- daedalus(country_x, "influenza_1918")

# with some infection parameters over-ridden by the user
output <- daedalus(
  "United Kingdom",
  daedalus_infection("influenza_1918", r0 = 1.3)
)

# with default initial conditions over-ridden by the user
output <- daedalus(
  "United Kingdom", "influenza_1918",
  initial_state_manual = list(p_infectious = 1e-3)
)

}
