% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus.R
\name{daedalus}
\alias{daedalus}
\title{DAEDALUS model for health, social and economic costs of a pandemic}
\usage{
daedalus(
  country,
  infection,
  response_strategy = c("none", "elimination", "economic_closures", "school_closures"),
  implementation_level = c("light", "heavy"),
  response_time = 30,
  response_threshold = NULL,
  vaccination_rate = 0,
  initial_state_manual = list(),
  time_end = 300,
  ...
)
}
\arguments{
\item{country}{A country or territory object of class \verb{<daedalus_country>},
\strong{or} a country or territory name from those included in the package;
see \link{country_names}.
Country-specific data such as the community and workplace contacts, the
demography, and the distribution of the workforce into economic sectors is
automatically accessed from package data for the relevant country name if it
is passed as a string.
To override package defaults for country characteristics, pass a
\verb{<daedalus_country>} object instead. See \code{\link[=daedalus_country]{daedalus_country()}} for more.}

\item{infection}{An infection parameter object of the class \verb{<infection>},
\strong{or} an epidemic name for which data are provided in the package; see
\link{epidemic_names} for parameters from a historical epidemic
or epidemic wave.
Passing the name as a string automatically accesses the default parameters
of an infection. Create an pass a \verb{<daedalus_infection>} to tweak infection
parameters.}

\item{response_strategy}{A string for the name of response strategy followed;
defaults to "none". The response strategy determines the country-specific
response threshold following which the response is activated. See
\code{response_threshold}.

While the response strategy is active, economic contacts are scaled using the
package data object \code{daedalus::closure_data}.}

\item{implementation_level}{A string for the level at which the strategy is
implemented; defaults to "light".}

\item{response_time}{A single numeric value for the time in days
at which the selected response is activated. This is ignored if the response
has already been activated by the hospitalisation threshold being reached.
Defaults to 30 days.}

\item{response_threshold}{A single numeric value for the total number of
hospitalisations that causes an epidemic response
(specified by \code{response_strategy}) to be triggered, if it has not already
been triggered via \code{response_time}. Currently defaults to \code{NULL}, and a
\code{country}-specific spare hospital capacity value is used from
\link{country_data}.
Pass a number to override the default country-specific threshold value.}

\item{vaccination_rate}{A single number in the range \link{0, 1} for the rate at
which individuals in all age groups are vaccinated. Defaults to 0.0 for no
vaccination active. A good value for an active vaccination strategy is
0.0025, or 0.25\% of each age group vaccinated per day, which corresponds to
average vaccination rates achieved during the Covid-19 pandemic.
This parameter will be heavily refactored in future package versions.}

\item{initial_state_manual}{An optional \strong{named} list with the names
\code{p_infectious} and \code{p_asymptomatic} for the proportion of infectious and
symptomatic individuals in each age group and economic sector.
Defaults to \code{1e-6} and \code{0.0} respectively.}

\item{time_end}{An integer-like value for the number of timesteps
at which to return data. This is treated as the number of days with data
returned for each day. Defaults to 300 days.}

\item{...}{Other arguments to be passed to the ODE solver; these are passed
to \code{\link[deSolve:ode]{deSolve::ode()}}.}
}
\value{
A \verb{<deSolve>} object.
}
\description{
Run the DAEDALUS model from R. This is a work in progress.
}
\details{
\subsection{Initial state}{

Users can pass the following initial state parameters to
\code{initial_state_manual}:
\itemize{
\item \code{p_infectious}: A single numeric value in the range \eqn{[0.0, 1.0]} giving
the proportion of individuals in each age group and economic sector that are
to be initialised as infectious. Defaults to \code{1e-6}, or one in every one
million as infectious.
\item \code{p_asymptomatic}: A single numeric value in the range \eqn{[0.0, 1.0]} for
the proportion of initially infectious individuals who are considered to be
asymptomatic. Defaults to 0.0.
}
}
}
\examples{
# country and infection specified by strings using default characteristics
output <- daedalus(
  "Canada", "influenza_1918"
)

# country passed as <daedalus_country> with some characteristics modified
country_x <- daedalus_country(
  "Canada",
  parameters = list(contact_matrix = matrix(5, 4, 4)) # uniform contacts
)
output <- daedalus(country_x, "influenza_1918")

# with some infection parameters over-ridden by the user
output <- daedalus(
  "United Kingdom",
  daedalus_infection("influenza_1918", r0 = 1.3)
)

# with default initial conditions over-ridden by the user
output <- daedalus(
  "United Kingdom", "influenza_1918",
  initial_state_manual = list(p_infectious = 1e-3)
)
}
