% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus.R
\name{daedalus}
\alias{daedalus}
\title{DAEDALUS model for health, social and economic costs of a pandemic}
\usage{
daedalus(
  country,
  epidemic,
  response_strategy = c("none", "elimination", "economic_closures", "school_closures"),
  implementation_level = c("light", "heavy"),
  response_threshold = NULL,
  end_threshold = NULL,
  country_params_manual = list(),
  infect_params_manual = list(),
  initial_state_manual = list(),
  time_end = 300
)
}
\arguments{
\item{country}{A country or territory name from among \code{country_names}.
Country-specific data such as the community and workplace contacts, the
demography, and the distribution of the workforce into economic sectors is
automatically accessed from package data for the relevant country name.
Some contact data values can be over-ridden by the user by passing them as a
named list to \code{country_params_manual}.}

\item{epidemic}{A string for the infection parameter set to use.
Infection parameter sets may be known by the name of the outbreak or the
causative pathogen;
see \strong{Details} for more on which epidemics are supported.
See \code{epidemic_names} for allowed names, as the name must match exactly.
Defaults to simulating an epidemic similar to Covid-19 outbreaks caused by
the wild type of the SARS-CoV-2 virus.

Selecting an epidemic automatically pulls in infection parameters
associated with the epidemic; these are stored as packaged data in
\code{infection_data}.
Default infection parameters for epidemics can be over-ridden by passing them
as a named list to \code{infection_params_manual}.}

\item{response_strategy}{A string for the name of response strategy followed;
defaults to "none". The response strategy determines how contacts are scaled
while the strategy is active, with contact scaling taken from the package
data object \code{daedalus::closure_data}.}

\item{implementation_level}{A string for the level at which the strategy is
implemented; defaults to "light".}

\item{response_threshold}{A named vector of length 1, with the value
representing the total number of individuals in a compartment, while the name
indicates the compartment name. The response chosen in \code{response_strategy} is
triggered when the response threshold is reached.}

\item{end_threshold}{A named vector of length 1, with the value
representing the total number of individuals in a compartment, while the name
indicates the compartment name. The response chosen in \code{response_strategy} is
ended when the end threshold is reached.}

\item{country_params_manual}{An optional \strong{named} list of country-specific
contact data. See \strong{Details} for allowed values.}

\item{infect_params_manual}{An optional \strong{named} list of infection
parameters that can be passed to over-ride the default values for the chosen
\code{epidemic}. See \strong{Details} for allowed values.}

\item{initial_state_manual}{An optional \strong{named} list with the names
\code{p_infectious} and \code{p_asymptomatic} for the proportion of infectious and
symptomatic individuals in each age group and economic sector.
Defaults to \code{1e-6} and \code{0.0} respectively.}

\item{time_end}{An integer-like value for the number of timesteps
at which to return data. This is treated as the number of days with data
returned for each day. Defaults to 300 days.}
}
\value{
A \verb{<deSolve>} object.
}
\description{
Run the DAEDALUS model from R. This is a work in progress.
}
\details{
\subsection{Included epidemics}{

Epidemics for which data are available are (pathogen in parentheses):
SARS 2004 (SARS-CoV-1), influenza 2009 (influenza A H1N1),
influenza 1957 (influenza A H2N2), influenza 1918 (influenza A H1N1),
Covid-19 wild type (SARS-Cov-2 wild type),
Covid-19 Omicron (SARS-CoV-2 omicron), Covid-19 Delta (SARS-CoV-2 delta).

DAEDALUS allows users to substitute default model parameters with custom
values by passing them in the appropriate argument (see examples).
These arguments are passed on to the internal functions
\code{make_country_parameters()} and \code{make_infection_parameters()} as
appropriate. Any arguments that are passed and are not suitable for those
functions will throw an error.
}

\subsection{Initial state}{

Users can pass the following initial state parameters to
\code{initial_state_manual}:
\itemize{
\item \code{p_infectious}: A single numeric value in the range \eqn{[0.0, 1.0]} giving
the proportion of individuals in each age group and economic sector that are
to be initialised as infectious. Defaults to \code{1e-6}, or one in every one
million as infectious.
\item \code{p_asymptomatic}: A single numeric value in the range \eqn{[0.0, 1.0]} for
the proportion of initially infectious individuals who are considered to be
asymptomatic. Defaults to 0.0.
}
}

\subsection{Country parameters}{

\code{country_params_manual} may be a named list of one or more of the following:
\itemize{
\item \code{contact_matrix}: A square numeric matrix with \code{N_AGE_GROUPS} (4) rows and
columns, giving the per-capita number of contacts between individuals of each
age group. By default, this is taken from \code{daedalus::country_data} based on
the user-specified \code{country}.
\item \code{contacts_workplace}: A numeric vector of length \code{N_ECON_SECTORS} (45),
giving the per-capita number of worker-to-worker interactions within economic
sectors. Defaults to a standard value provided by the package as
\code{daedalus::economic_contacts}.
\item \code{contacts_consumer_worker}: A numeric matrix with \code{N_ECON_SECTORS} rows
(45) and \code{N_AGE_GROUPS} columns (4) giving the per-capita contacts between
individuals in each economic sector and consumers. Defaults to the
sector-wise workplace contacts distributed in proportion to country
demography.
\item \code{contacts_between_sectors}: A square numeric matrix with \code{N_ECON_SECTORS}
(45) rows and columns and with a zero diagonal, giving the number of
worker-to-worker contacts across economic sectors. Defaults to the package
standard of a null matrix as between-sector-contacts are not currently
modelled.
}
}

\subsection{Infection parameters}{

\code{infect_params_manual} may be a named list of one or more of the following:
\itemize{
\item \code{r0}: A single numeric value for the basic reproduction value of the
infection \eqn{R_0}.
\item \code{sigma}: A single numeric value > 0.0 for the rate of transition from the
exposed compartment to one of two infectious compartments.
\item \code{p_sigma}: A single numeric value in the range \eqn{[0.0, 1.0]} for the
proportion of infectious individuals who are also symptomatic. Asymptomatic
individuals can have a different contribution to the force of infection from
symptomatic individuals.
\item \code{epsilon}: A single numeric value for the relative contribution of
asymptomatic infectious individuals to the force of infection (compared to
symptomatic individuals).
\item \code{gamma_Is}: A single numeric value for the recovery rate of infectious
individuals who are not hospitalised.
\item \code{gamma_Ia}: A single numeric value for the recovery rate from asymptomatic
infection.
\item \code{gamma_H}: A numeric vector of length \code{N_AGE_GROUPS} (4) for the
age-specific recovery rate for individuals who are hospitalised.
\item \code{eta}: A numeric vector of length \code{N_AGE_GROUPS} (4) for the age-specific
hospitalisation rate for individuals who are infectious and symptomatic.
\item \code{omega}: A numeric vector of length \code{N_AGE_GROUPS} (4) for the age-specific
mortality rate for individuals who are hospitalised.
\item \code{rho}: A single numeric value for the rate at which infection-derived
immunity wanes, returning individuals in the 'recovered' compartment to the
'susceptible' compartment.
}
}
}
\examples{
# with default infection parameters associated with an epidemic
output <- daedalus(country = "United Kingdom", epidemic = "influenza_1918")

# with some infection parameters over-ridden by the user
output <- daedalus(
  country = "United Kingdom", epidemic = "influenza_1918",
  infect_params_manual = list(r0 = 1.3)
)

# with some country parameters over-ridden by the user
output <- daedalus(
  country = "United Kingdom", epidemic = "influenza_1918",
  country_params_manual = list(contact_matrix = matrix(1, 4, 4))
)

# with default initial conditions over-ridden by the user
output <- daedalus(
  country = "United Kingdom", epidemic = "influenza_1918",
  initial_state_manual = list(p_infectious = 1e-3)
)
}
