% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus2.R
\name{daedalus2}
\alias{daedalus2}
\title{DAEDALUS model implemented with dust}
\usage{
daedalus2(
  country,
  infection,
  response_strategy = NULL,
  vaccine_investment = NULL,
  response_time = 30,
  response_duration = 365,
  initial_state_manual = NULL,
  time_end = 600,
  ...
)
}
\arguments{
\item{country}{A country or territory object of class \verb{<daedalus_country>},
\strong{or} a country or territory name from those included in the package;
see \link{country_names}, \strong{or} a country ISO2 or ISO3 code; see
\link{country_codes_iso2c} and \link{country_codes_iso3c} .
Country-specific data such as the community and workplace contacts, the
demography, and the distribution of the workforce into economic sectors is
automatically accessed from package data for the relevant country name if it
is passed as a string.
To override package defaults for country characteristics, pass a
\verb{<daedalus_country>} object instead. See \code{\link[=daedalus_country]{daedalus_country()}} for more.}

\item{infection}{An infection parameter object of the class
\verb{<daedalus_infection>}, \strong{or} an epidemic name for which data are provided
in the package; see \link{epidemic_names} for historical epidemics
or epidemic waves for which parameters are available.

Passing the name as a string automatically accesses the default parameters
of an infection. Create and pass a \verb{<daedalus_infection>} to tweak infection
parameters using \code{\link[=daedalus_infection]{daedalus_infection()}}.}

\item{response_strategy}{A string for the name of response strategy followed;
defaults to "none". The response strategy determines the country-specific
response threshold following which the response is activated. See
\code{response_threshold}.

While the response strategy is active, economic contacts are scaled using the
package data object \code{daedalus::closure_data}.}

\item{vaccine_investment}{Either a single string or a
\verb{<daedalus_vaccination>} object specifying the vaccination parameters
associated with an advance vaccine-investment scenario. Defaults to \code{"none"},
for no advance vaccine investment. In this case, vaccination begins 365 days
(1 year) after the simulation begins, at a low rate across all age groups.
Other accepted values are \code{"low"}, \code{"medium"} and \code{"high"}. See
\code{\link[=daedalus_vaccination]{daedalus_vaccination()}} for more information.}

\item{response_time}{A single numeric value for the time in days
at which the selected response is activated. This is ignored if the response
has already been activated by the hospitalisation threshold being reached.
Defaults to 30 days.}

\item{response_duration}{A single integer-ish number that gives the number of
days after \code{response_time} that an NPI should end.}

\item{initial_state_manual}{An optional \strong{named} list with the names
\code{p_infectious} and \code{p_asymptomatic} for the proportion of infectious and
symptomatic individuals in each age group and economic sector.
Defaults to \code{1e-6} and \code{0.0} respectively.}

\item{time_end}{An integer-like value for the number of timesteps
at which to return data. This is treated as the number of days with data
returned for each day. Defaults to 300 days.}

\item{...}{Optional arguments that are passed to \code{\link[dust2:dust_ode_control]{dust2::dust_ode_control()}}.}
}
\description{
\code{daedalus::daedalus2()} will eventually replace \code{daedalus::daedalus()} with
the model implemented in C++ to integrate with \pkg{dust2}
(and in future \pkg{dust}). \emph{This is a work in progress!}
}
\details{
\strong{Note that} \code{daedalus2()} currently uses a default vaccination strategy of
\emph{no vaccination}, using the internal helper function \code{\link[=dummy_vaccination]{dummy_vaccination()}}.
This is expected to become the default for \code{\link[=daedalus]{daedalus()}} when it is replaced
with \code{daedalus2()}.

\strong{Note also} that \code{daedalus2()} vaccination begins at the model start time,
and not at the time specified in \code{vaccine_investment}. This functionality
will be updated soon.
}
\examples{
# not yet ready for conversion to `<daedalus_output>`
output <- daedalus2("GBR", "sars_cov_1")

names(output)
}
