% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus2.R
\name{daedalus2}
\alias{daedalus2}
\title{DAEDALUS model implemented with dust}
\usage{
daedalus2(country, infection, time_end = 100)
}
\arguments{
\item{country}{A country or territory object of class \verb{<daedalus_country>},
\strong{or} a country or territory name from those included in the package;
see \link{country_names}, \strong{or} a country ISO2 or ISO3 code; see
\link{country_codes_iso2c} and \link{country_codes_iso3c} .
Country-specific data such as the community and workplace contacts, the
demography, and the distribution of the workforce into economic sectors is
automatically accessed from package data for the relevant country name if it
is passed as a string.
To override package defaults for country characteristics, pass a
\verb{<daedalus_country>} object instead. See \code{\link[=daedalus_country]{daedalus_country()}} for more.}

\item{infection}{An infection parameter object of the class \verb{<infection>},
\strong{or} an epidemic name for which data are provided in the package; see
\link{epidemic_names} for parameters from a historical epidemic
or epidemic wave.
Passing the name as a string automatically accesses the default parameters
of an infection. Create an pass a \verb{<daedalus_infection>} to tweak infection
parameters.}

\item{time_end}{An integer-like value for the number of timesteps
at which to return data. This is treated as the number of days with data
returned for each day. Defaults to 300 days.}
}
\description{
\code{daedalus::daedalus2()} will eventually replace \code{daedalus::daedalus()} with
the model implemented in C++ to integrate with \pkg{dust2}
(and in future \pkg{dust}). \emph{This is a work in progress!}
}
\examples{
daedalus2("GBR", "sars_cov_1")
}
