% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R, R/pkg_generics.R
\name{daedalus_costs}
\alias{daedalus_costs}
\alias{get_costs.daedalus_output}
\alias{get_costs.data.frame}
\alias{get_costs}
\title{Get epidemic losses from a DAEDALUS model run}
\usage{
\method{get_costs}{daedalus_output}(
  x,
  summarise_as = c("none", "total", "domain"),
  productivity_loss_infection = 1,
  ...
)

\method{get_costs}{data.frame}(
  x,
  comp_non_working,
  comp_infected,
  comp_dead,
  workforce,
  daily_gva,
  productivity_loss_infection,
  vsl_by_age,
  life_expectancy,
  value_school_year,
  n_students,
  edu_effectiveness_remote,
  npi_data = NULL,
  summarise_as = c("none", "total", "domain"),
  ...
)

get_costs(x, ...)
}
\arguments{
\item{x}{A \verb{<daedalus_output>} object from a call to \code{\link[=daedalus]{daedalus()}}, or a
\verb{<data.frame>} of an epidemic timeseries.

If \code{x} is a \verb{<data.frame>}, it must have the columns \code{"time"},
\code{"compartment"}, \code{"age_group"}, \code{"econ_sector"} and \code{"value"}, giving the
model time, the epidemiological compartment, and the number of individuals
of each age group and economic sector in each compartment at each time.
See \strong{Details} for more on the expectations around how this dataset is
organised.}

\item{summarise_as}{A string from among "none", "total", or "domain", for how
the costs should be returned. Select "none", the default, for the raw costs
along with overall and domain-specific totals; "total" for the overall cost,
and "domain" for the total costs per domain; the domains are 'economic',
'education', and 'life years'.}

\item{productivity_loss_infection}{A single number in the range
\eqn{[0, 1]} giving the loss in productivity associated with symptomatic
infection. Currently defaults to 1.0 for compatibility with earlier function
versions.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Other arguments to class
methods.}

\item{comp_non_working}{A character vector giving the names of compartments
in which individuals are assumed to have a reduced ability to work. This
may (and should) include compartments where individuals are infected and
symptomatic, as well as hospitalised or dead.}

\item{comp_infected}{A character vector giving the name of compartments in
which individuals are assumed to be infected with reduced ability to work.
This should \strong{not} include compartments where individuals are assumed to
have \strong{no} ability to work (hospitalised or dead).}

\item{comp_dead}{A string giving the name of the compartment holding the
count of deaths. Note that only a single deaths compartment is supported, so
models with multiple death states should pass the \strong{final state}. Used to
calculate the value of lives lost.}

\item{workforce}{A numeric vector of the number of workers in each economic
sector of the population. Must be absolute numbers (i.e., not scaled to
thousands or millions). Must have the same length as \code{daily_gva}.}

\item{daily_gva}{A numeric vector of the daily gross-value added (GVA)
in million dollars, by each economic sector of the population (or territory
or state) in which the epidemic occurs.}

\item{vsl_by_age}{A numeric vector of the value of a statistical life (VSL)
for each age age group in \code{x}.}

\item{life_expectancy}{A numeric vector of the life-expectancy per age group.
Expected to be the same length as \code{vsl_by_age}.}

\item{value_school_year}{A single number giving the value of a school year.
Users can use the helper function \code{\link[=get_value_school_year]{get_value_school_year()}} if a population's
gross income per-capita is known to apply the internal Daedalus model for
valuing a school year, or a value obtained in some other way.}

\item{n_students}{A single number for the number of students in the
population, whose education is expected to be impacted by the epidemic.}

\item{edu_effectiveness_remote}{A single number for the effectiveness of
remote education, in the range \eqn{[0, 1]}.}

\item{npi_data}{A list with numeric elements that mirrors the elements of a
\verb{<daedalus_npi>}, with the following names:
\itemize{
\item \code{npi_times_start}: A numeric vector of the start times of any NPIs
modelled.
\item \code{npi_times_end}: A numeric vector of the end times of modelled NPIs.
\item \code{npi_durations}: A numeric vector of the durations of NPIs.
\item \code{npi_periods}: A numeric vector giving integer-ish elements in the ranges
specified by the start and end times.
\item \code{openness}: Either a numeric vector (if modelling a single NPI), or a list
of numeric vectors (if modelling multiple NPIs), giving the openness
coefficients of NPI regimes. Vector lengths must be the same as \code{workforce}.
}}
}
\value{
A list of different cost values, including the total cost. See
\strong{Details} for more information.
}
\description{
Calculate the costs of an epidemic and any mitigation measures, either from a
Daedalus model run, or from a timeseries of epidemic data and relevant
demographic and economic parameters.
}
\details{
\subsection{Expectations for non-Daedalus model data}{

When a \verb{<data.frame> x} is passed, it must have at least one row, at least
some non-missing values, and:
\enumerate{
\item Columns \code{"time"}, \code{"age_group"}, \code{"compartment"}, \code{"econ_sector"}, and
\code{"value"}, giving the model time (numeric), age group identifier (character),
epidemiological compartment identifier (character), economic sector
identifier (character), and the value of the number of individuals in each
age, epi-compartment, and economic group at each time (numeric). Numeric
values must be non-negative \eqn{> 0}, non-missing, and finite.
\item Compartments in \code{comp_infected}, \code{comp_non_working}, and \code{comp_dead} must
be found in \code{x$compartment}.
\item All compartments of \code{comp_infected} and \code{comp_dead} must be in
\code{comp_non_working}.
}
}

\subsection{Output}{

The total cost in million dollars is returned as \code{total_cost}. This is
comprised of the following costs.
\subsection{Economic costs}{

A three element list of \code{economic_cost_total}, the total costs from pandemic
impacts on economic sectors, including both costs of lost gross value added
(GVA) due to pandemic-control restrictions or closures
(\code{economic_cost_closures}), and pandemic-related absences due to illness and
death (\code{economic_cost_absences}).
}

\subsection{Educational costs}{

A three element list of \code{education_cost_total}, the total costs from pandemic
impacts on education due to pandemic-control restrictions or closures
(\code{education_cost_closures}), and pandemic-related absences due to illness and
death (\code{education_cost_absences}).
}

\subsection{Life-value lost}{

A four-element vector (for the number of age groups) giving the value of
life-years lost per age group. This is calculated as the life-expectancy of
each age group times the value of a statistical life, with all years assumed
to have the same value.
}

\subsection{Life-years lost}{

A four-element vector (for the number of age groups) giving the value of
life-years lost per age group. This is calculated as the life-expectancy of
each age group times the number of deaths in that age group. No quality
adjustment is applied.
}

}
}
\examples{
# for <daedalus_output> objects from `daedalus()`
o <- daedalus("CAN", "sars_cov_1", time_end = 100)
get_costs(o, "domain")

# for a data.frame of epi compartment timeseries, as from any epi model
output <- daedalus("Canada", "influenza_1918", time_end = 100)

data <- get_data(output)
comp_non_working <- c(
  "infect_symp",
  "hospitalised_recov",
  "hospitalised_death",
  "dead"
)
comp_infected <- "infect_symp"
comp_dead <- "dead"

daily_gva <- output$country$gva
workforce <- output$country$workers
vsl_by_age <- output$country$vsl
life_expectancy <- output$country$life_expectancy

value_school_year <- 1e6 # 1 million dollars
n_students <- output$country$demography[3L]

edu_effectiveness_remote <- 0.33

productivity_loss_infection <- 1.0

get_costs(
  data,
  comp_non_working,
  comp_infected,
  comp_dead,
  daily_gva,
  workforce,
  vsl_by_age,
  life_expectancy,
  value_school_year,
  n_students,
  productivity_loss_infection = productivity_loss_infection,
  edu_effectiveness_remote = edu_effectiveness_remote,
  summarise_as = "domain"
)

}
