% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus_multi.R
\name{daedalus_multi_infection}
\alias{daedalus_multi_infection}
\title{Run daedalus for multiple infection parameter sets}
\usage{
daedalus_multi_infection(
  country,
  infection,
  response_strategy = NULL,
  vaccine_investment = NULL,
  response_time = 30,
  auto_social_distancing = c("off", "independent", "npi_linked"),
  initial_state_manual = NULL,
  time_end = 600,
  ...
)
}
\arguments{
\item{country}{A country or territory object of class \verb{<daedalus_country>},
\strong{or} a country or territory name from those included in the package;
see \link[daedalus.data:country_names_codes]{daedalus.data::country_names}, \strong{or} a country ISO2 or ISO3 code; see
\link[daedalus.data:country_names_codes]{daedalus.data::country_codes_iso2c} and
\link[daedalus.data:country_names_codes]{daedalus.data::country_codes_iso3c}.
Country-specific data such as the community and workplace contacts, the
demography, and the distribution of the workforce into economic sectors is
automatically accessed from package data for the relevant country name if it
is passed as a string.
To override package defaults for country characteristics, pass a
\verb{<daedalus_country>} object instead. See \code{\link[=daedalus_country]{daedalus_country()}} for more.}

\item{infection}{A list of \verb{<daedalus_infection>} objects.}

\item{response_strategy}{A string for the name of response strategy followed,
a numeric of length 45 (number of economic sectors), or a \verb{<daedalus_npi>}
object. Defaults to "none".
While the response strategy is active, economic contacts are scaled using the
package data object \code{daedalus.data::closure_strategy_data}.}

\item{vaccine_investment}{Either a single string or a
\verb{<daedalus_vaccination>} object specifying the vaccination parameters
associated with an advance vaccine-investment scenario. Defaults to \code{NULL}
for absolutely no vaccination in the model.
A vaccination investment of \code{"none"} indicates no \emph{prior} investment, but
the model will include vaccination beginning after 1 year, at a low rate
across all age groups. Other accepted values are \code{"low"}, \code{"medium"} and
\code{"high"}. See \code{\link[=daedalus_vaccination]{daedalus_vaccination()}} for more information.}

\item{response_time}{A single numeric value for the time in days
at which the selected response is activated. This is ignored if the response
has already been activated by the hospitalisation threshold being reached.
Defaults to 30 days. Responses have a default maximum duration of 365 days.
This can be changed by passing a \verb{<daedalus_npi>} object to
\code{response_strategy}.}

\item{auto_social_distancing}{A string giving the option for the form of
spontaneous social distancing in the model, which reduces infection
transmission as a function of daily deaths. See \strong{Details} for more.}

\item{initial_state_manual}{An optional \strong{named} list with the names
\code{p_infectious}, \code{p_asymptomatic}, and \code{p_immune}.
\code{p_infectious} and \code{p_asymptomatic} give the proportion of
infectious and symptomatic individuals in each age group and economic sector.
Defaults to \code{1e-6} and \code{0.0} respectively.
\code{p_immune} may be a single number in the range \verb{0.0 <= p_immune <= 1.0} or a
4-element vector in that range (the number of age groups in the model), for
the proportion of individuals in the population or in each age group that
have some pre-existing immunity to infection (reduced susceptibility). See
\strong{Details} for more.}

\item{time_end}{An integer-like value for the number of timesteps
at which to return data. This is treated as the number of days with data
returned for each day. Defaults to 300 days.}

\item{...}{Optional arguments that are passed to \code{\link[dust2:dust_ode_control]{dust2::dust_ode_control()}}.}
}
\value{
A list of \verb{<daedalus_output>} objects.
}
\description{
Run \code{\link[=daedalus]{daedalus()}} for multiple parameter sets, with the intended use case of
running the model under uncertainty in infection parameters.
}
\details{
See details of how \code{auto_social_distancing} affects the model in
documentation for \code{\link[=daedalus]{daedalus()}}.
}
