% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus_rtm.R
\name{daedalus_rtm}
\alias{daedalus_rtm}
\title{DAEDALUS model for health, social and economic costs of a pandemic}
\usage{
daedalus_rtm(
  country,
  infection,
  response_strategy = NULL,
  response_time_start = NULL,
  response_time_end = NULL,
  initial_state_manual = list(),
  time_end = 300,
  ...
)
}
\arguments{
\item{country}{A country or territory object of class \verb{<daedalus_country>},
\strong{or} a country or territory name from those included in the package;
see \link{country_names}, \strong{or} a country ISO2 or ISO3 code; see
\link{country_codes_iso2c} and \link{country_codes_iso3c} .
Country-specific data such as the community and workplace contacts, the
demography, and the distribution of the workforce into economic sectors is
automatically accessed from package data for the relevant country name if it
is passed as a string.
To override package defaults for country characteristics, pass a
\verb{<daedalus_country>} object instead. See \code{\link[=daedalus_country]{daedalus_country()}} for more.}

\item{infection}{An infection parameter object of the class \verb{<infection>},
\strong{or} an epidemic name for which data are provided in the package; see
\link{epidemic_names} for parameters from a historical epidemic
or epidemic wave.
Passing the name as a string automatically accesses the default parameters
of an infection. Create an pass a \verb{<daedalus_infection>} to tweak infection
parameters.}

\item{response_strategy}{A string for the name of response strategy followed;
defaults to "none". The response strategy determines the country-specific
response threshold following which the response is activated. See
\code{response_threshold}.

While the response strategy is active, economic contacts are scaled using the
package data object \code{daedalus::closure_data}.}

\item{initial_state_manual}{An optional \strong{named} list with the names
\code{p_infectious} and \code{p_asymptomatic} for the proportion of infectious and
symptomatic individuals in each age group and economic sector.
Defaults to \code{1e-6} and \code{0.0} respectively.}

\item{time_end}{An integer-like value for the number of timesteps
at which to return data. This is treated as the number of days with data
returned for each day. Defaults to 300 days.}

\item{...}{Other arguments to be passed to the ODE solver; these are passed
to \code{\link[deSolve:ode]{deSolve::ode()}}.}
}
\value{
A \verb{<deSolve>} object.
}
\description{
Run the DAEDALUS model from R. This is a work in progress.
}
\details{
\subsection{Initial state}{

Users can pass the following initial state parameters to
\code{initial_state_manual}:
\itemize{
\item \code{p_infectious}: A single numeric value in the range \eqn{[0.0, 1.0]} giving
the proportion of individuals in each age group and economic sector that are
to be initialised as infectious. Defaults to \code{1e-6}, or one in every one
million as infectious.
\item \code{p_asymptomatic}: A single numeric value in the range \eqn{[0.0, 1.0]} for
the proportion of initially infectious individuals who are considered to be
asymptomatic. Defaults to 0.0.
}
}
}
\examples{
# country and infection specified by strings using default characteristics
output <- daedalus(
  "Canada", "influenza_1918"
)

# country passed as <daedalus_country> with some characteristics modified
country_x <- daedalus_country(
  "Canada",
  parameters = list(contact_matrix = matrix(5, 4, 4)) # uniform contacts
)
output <- daedalus(country_x, "influenza_1918")

# with some infection parameters over-ridden by the user
output <- daedalus(
  "United Kingdom",
  daedalus_infection("influenza_1918", r0 = 1.3)
)

# with default initial conditions over-ridden by the user
output <- daedalus(
  "United Kingdom", "influenza_1918",
  initial_state_manual = list(p_infectious = 1e-3)
)
}
