% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_vaccination.R
\name{dummy_vaccination}
\alias{dummy_vaccination}
\title{Dummy vaccination}
\usage{
dummy_vaccination()
}
\value{
A \code{daedalus_vaccination} object intended to have no effect;
vaccination rate and efficacy are set to zero.
}
\description{
Dummy vaccination
}
\keyword{internal}
