% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{econ_constants}
\alias{econ_constants}
\alias{edu_effectiveness_remote}
\alias{edu_annual_ror}
\alias{work_expected_years}
\alias{earnings_loss_discount}
\title{Economic constants used in DAEDALUS}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
edu_effectiveness_remote

edu_annual_ror

work_expected_years

earnings_loss_discount
}
\description{
Economic constants used in DAEDALUS
}
\details{
The Daedalus model uses these constants for economic value calculations:
\enumerate{
\item \code{edu_effectiveness_remote}: The effectiveness of remote education.
\item \code{edu_annual_ror}: The rate of return on a year of education.
\item \code{work_expected_years}: The expected number of years of work for school-age
children.
\item \code{earnings_loss_discount}: The discounting rate used in the calculation of
lost earnings.
}
}
\keyword{econ_constant}
