% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{epi_constants}
\alias{epi_constants}
\alias{COMPARTMENTS}
\alias{N_MODEL_COMPARTMENTS}
\alias{i_S}
\title{Epidemiological compartments and indices}
\format{
An object of class \code{character} of length 10.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.
}
\usage{
COMPARTMENTS

N_MODEL_COMPARTMENTS

i_S
}
\value{
\code{COMPARTMENTS} returns a character vector of the epidemiological
compartment names.

All other constants return integer values.
}
\description{
Names and indices for the epidemiological compartments used in
DAEDALUS, for reuse in model code.
}
\details{
DAEDALUS has 7 epidemiological compartments: susceptible, exposed, infectious
and symptomatic ("infect_symp"), infectious and asymptomatic ("infect_asymp")
, hospitalised, recovered, and dead.

There are 3 additional compartments that track the number of new infections,
new hospitalisations, and new deaths (\code{i_dE}, \code{i_dH}, and \code{i_dD}).
}
\keyword{epi_constant}
