% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{epi_constants}
\alias{epi_constants}
\alias{COMPARTMENTS}
\alias{i_S}
\alias{i_E}
\alias{i_Is}
\alias{i_Ia}
\alias{i_H}
\alias{i_R}
\alias{i_D}
\alias{N_EPI_COMPARTMENTS}
\title{Epidemiological compartments and indices}
\format{
An object of class \code{character} of length 7.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.
}
\usage{
COMPARTMENTS

i_S

i_E

i_Is

i_Ia

i_H

i_R

i_D

N_EPI_COMPARTMENTS
}
\value{
\code{COMPARTMENTS} returns a character vector of the epidemiological
compartment names.

All other constants return integer values.
}
\description{
Names and indices for the epidemiological compartments used in
DAEDALUS, for reuse in model code.
}
\details{
DAEDALUS has 7 epidemiological compartments: susceptible, exposed, infectious
and symptomatic ("infect_symp"), infectious and asymptomatic ("infect_asymp")
, hospitalised, recovered, and dead.
}
\keyword{datasets}
\keyword{internal}
