% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_helpers.R
\name{epi_output_helpers}
\alias{epi_output_helpers}
\alias{get_incidence}
\alias{get_epidemic_summary}
\title{Calculate daily incidences and summarise epidemic measures}
\usage{
get_incidence(
  data,
  measures = c("infections", "hospitalisations", "deaths"),
  groups = NULL
)

get_epidemic_summary(
  data,
  measures = c("infections", "hospitalisations", "deaths"),
  groups = NULL
)
}
\arguments{
\item{data}{Either a \verb{<data.frame>} or a \verb{<daedalus_output>} object.}

\item{measures}{A character vector of one or more of the following:
\code{"epidemic_size"}, \code{"total_hospitalisations"} or \code{"total_deaths"} for the
measure to return.
Defaults to returning all three in long format.}

\item{groups}{An optional character vector of grouping variables that
correspond to model strata. Defaults to \code{NULL} which gives incidence across
the whole population. Allowed groups correspond to modelled strata:
\code{"age_group"}, \code{"vaccine_group"}, and \code{"econ_sector"}.}
}
\value{
A \verb{<data.frame>} in long format, with at least one entry per
model timestep, measure, and group chosen.
\itemize{
\item \code{get_incidence()} returns a data frame with the number of daily new
infections, new hospitalisations, and/or new deaths in each of the groups
specified by \code{groups}.
\item \code{get_epidemic_summary()} returns a data frame with the measure specified
in \code{measure} for each of the groups specified by \code{groups}.
}
}
\description{
Functions to quickly summarise timeseries data from \code{daedalus()}
while allowing grouping by different strata.
}
\examples{
data <- daedalus("Canada", "sars_cov_1")

# new infections
new_infections <- get_incidence(data, "infections")

# epidemic summary
get_epidemic_summary(
  data,
  measures = c("epidemic_size", "total_deaths", "total_hospitalisations"),
  groups = "age_group"
)

}
