% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus.R
\name{get_daedalus_response_times}
\alias{get_daedalus_response_times}
\title{Get model response times from dust2 output}
\usage{
get_daedalus_response_times(event_data, time_end)
}
\arguments{
\item{time_end}{The model end time, passed from \code{\link[=daedalus]{daedalus()}}.}

\item{output}{dust2 output from \code{daedalus_internal()}.}
}
\value{
A vector of event start and end times suitable for a
\verb{<daedalus_output>} object. Returns model end time if there is no response
end time.
}
\description{
Get model response times from dust2 output
}
\keyword{internal}
