% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_country.R, R/pkg_generics.R
\name{get_data}
\alias{get_data}
\alias{get_data.country}
\title{Get \verb{<country>} parameters}
\usage{
\method{get_data}{country}(x, ...)

get_data(x, ...)
}
\arguments{
\item{x}{An S3 class object from the \pkg{daedalus} package. Currently only
\verb{<country>} objects are supported.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Strings giving the names of
elements to be accessed.}
}
\value{
If only a single parameter name is passed in \code{...}, returns an object
of the appropriate class (e.g. numeric vector for \code{"demography"}).

If multiple parameters are accessed at once, returns a named list of the
requested parameters.
}
\description{
Generic and methods for S3 classes for safely getting class
parameters.
}
\examples{
# simple example of getting data
country_A <- country("United Kingdom")
get_data(country_A, "demography")

get_data(country_A, "demography", "contact_matrix")
}
