% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_country.R, R/class_infection.R,
%   R/pkg_generics.R
\name{get_data}
\alias{get_data}
\alias{get_data.daedalus_country}
\alias{get_data.daedalus_infection}
\title{Get parameters from DAEDALUS classes}
\usage{
\method{get_data}{daedalus_country}(x, ...)

\method{get_data}{daedalus_infection}(x, ...)

get_data(x, ...)
}
\arguments{
\item{x}{An S3 class object from the \pkg{daedalus} package of the
\verb{<daedalus_country>} or \verb{<infection>} class.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Strings giving the names of
elements to be accessed.}
}
\value{
If only a single parameter name is passed in \code{...}, returns an object
of the appropriate class (e.g. numeric vector for \code{"demography"}).

If multiple parameters are accessed at once, returns a named list of the
requested parameters.
}
\description{
Generic and methods for S3 classes for safely getting class
parameters.
}
\examples{
# simple example of getting data
country_A <- daedalus_country("United Kingdom")
get_data(country_A, "demography")

get_data(country_A, "demography", "contact_matrix")

disease_x <- daedalus_infection("sars_cov_1", r0 = 1.9)
get_data(disease_x, "r0")
}
