% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{get_fiscal_costs}
\alias{get_fiscal_costs}
\title{Get pandemic fiscal costs from a model run}
\usage{
get_fiscal_costs(
  x,
  support_level = 0.33,
  price_vax = 1,
  price_npi = 1,
  uptake_npi = 1,
  interest_rate = 4,
  tax_rate = 25,
  spending_rate = 45
)
}
\arguments{
\item{x}{A \verb{<daedalus_output>} object.}

\item{support_level}{The proportion of pandemic-related economic losses that
a government compensates, as a proportion.}

\item{price_vax}{The per-dose price of vaccination.}

\item{price_npi}{The per-day and per-person price of implementing any
pandemic response. May include costs such as testing or masks.}

\item{uptake_npi}{The extent to which NPIs are taken up by the population;
essentially a number that modifies (reduces) the total cost of implementing
an NPI.}

\item{interest_rate}{The annual interest rate on government borrowing for
pandemic response.}

\item{tax_rate}{The annual mean tax rate on economic output; used to
calculate government revenue.}

\item{spending_rate}{The annual mean rate of public spending as a percentage
of GDP.}
}
\value{
A two-element list giving:
\itemize{
\item A list of \code{fiscal_costs} with elements giving the daily fiscal costs
\eqn{\text{TCG_t}} and a breakdown of these costs, as well as interest;
\item A time-series \code{daily_public_debt} of the public debt \eqn{b_t}, which is
the net of baseline public spending, pandemic response costs, and
pandemic-impacted revenue.
}
}
\description{
A helper function that post-processes a \verb{<daedalus_output>} object to
calculate the costs to a national government (fiscal costs) of responding to
a pandemic. Includes costs of economic support, vaccinations given, and NPIs
administered or implemented.
}
\details{
Default argument values are intended to be representative. This
function is intended to be called after \code{\link[=daedalus]{daedalus()}} and parameters
required to compute fiscal costs may need to be bundled along with model
outputs.
Note that all rates (interest rate, spending rate, and tax rate) are given as
annual percentages, and are simply divided by 365 to give the daily
rate.
}
\examples{
# get fiscal costs for UK with SARS-CoV-2
# and both closures and vaccinations
o <- daedalus("GBR", "sars_cov_2_pre_alpha", "economic_closures", "high")
get_fiscal_costs(o)

# also works when no closures are applied
o <- daedalus("CAN", "influenza_2009")
get_fiscal_costs(o)
}
