% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{get_fiscal_costs}
\alias{get_fiscal_costs}
\title{Get pandemic fiscal costs from a model run}
\usage{
get_fiscal_costs(
  x,
  support_level = 0.2,
  price_vax = 1,
  price_npi = 1,
  uptake_npi = 1,
  interest_rate = 4,
  tax_rate = 35,
  spending_rate = 45,
  starting_debt = 0
)
}
\arguments{
\item{x}{A \verb{<daedalus_output>} object.}

\item{support_level}{The proportion of pandemic-related economic losses that
a government compensates, as a proportion.}

\item{price_vax}{The per-dose price of vaccination.}

\item{price_npi}{The per-day and per-person price of implementing any
pandemic response. May include costs such as testing or masks.}

\item{uptake_npi}{The extent to which NPIs are taken up by the population;
essentially a number that modifies (reduces) the total cost of implementing
an NPI.}

\item{interest_rate}{The annual interest rate on government borrowing for
pandemic response.}

\item{tax_rate}{The annual mean tax rate on economic output; used to
calculate government revenue.}

\item{spending_rate}{The annual mean rate of public spending as a percentage
of GDP.}

\item{starting_debt}{The value of national starting debt. Currently assumed
to be zero while country-specific data are collected.}
}
\value{
A two-element list giving:
\itemize{
\item A list of \code{fiscal_costs} with elements giving the fiscal costs
\eqn{\text{TCG_t}} and a breakdown of these costs, as well as interest;
\item A list of \code{public_debt} of the public debt \eqn{b_t}, which is
the net of baseline public spending, pandemic response costs, and
pandemic-impacted revenue.
}
}
\description{
A helper function that post-processes a \verb{<daedalus_output>} object to
calculate the costs to a national government (fiscal costs) of responding to
a pandemic. Includes costs of economic support, vaccinations given, and NPIs
administered or implemented.
}
\details{
Default argument values are intended to be representative. This
function is intended to be called after \code{\link[=daedalus]{daedalus()}} and parameters
required to compute fiscal costs may need to be bundled along with model
outputs.
Note that all rates (interest rate, spending rate, and tax rate) are given as
annual percentages, and are simply divided by 365 to give the daily
rate.
\subsection{Public spending}{

Spending is calculated as:
\eqn{G_t = \sigma(\text{GVA}^* - \text{GVA}_t) + Cv_t + Cp_t}

where \eqn{Cv_t} is the time-specific cost of vaccination, and is calculated
as the cost of new vaccinations in each timestep: \eqn{P_v \times \delta V_t}
.

\eqn{Cp_t} is the time-specific cost of implementing pandemic response,
and is calculated as \eqn{\psi (\bar N - D_t) P_p}, where
\eqn{\bar N - D_t} is the remaining number of individuals in the population,
and \eqn{\psi} is the proportion taking up any protection offered by the
response.

The total cost to the public of the pandemic over each day of the pandemic
is assumed to be made up of borrowing, and is then
\eqn{\text{TCG}_t = G_t + (1 + R_t^T) \text{TCG}_{t - 1}}, where
\eqn{R_t^T} is the daily rate of interest to be paid on the borrowed amount.
This rate is currently modelled as being constant over time.

The total public debt at the end of the pandemic \eqn{b_t} is then the sum of
total public spending on the pandemic \eqn{\text{TCG}_t},
existing day-to-day public spending \eqn{\bar G} which is assumed to be a
fraction of daily GDP \eqn{\nu \text{GDP}}, existing debt owed due to past
daily spending, interest to be paid on the debt, less the revenues collected
\eqn{\mu \text{GVA}_{t-1}}. The daily GVA is the pre-pandemic GVA scaled by
the available labour supply during the pandemic, taking into account labour
restrictions due to illness-related absences and deaths, and response-related
restrictions.

GDP is calculated as the sum of sector-specific daily GVA, and existing debt
is currently assumed to be zero and is not included in the equation.

\eqn{b_t = \bar G + \text{TCG}_{t-1} + (1 + R_{t-1}^T) b_{t-1} -
  \mu \text{GVA}_{t-1}
}
}
}
\examples{
# get fiscal costs for UK with SARS-CoV-2
# and both closures and vaccinations
o <- daedalus(
  "GBR", "sars_cov_2_pre_alpha",
  "economic_closures", "high",
  time_end = 30
)
fc <- get_fiscal_costs(o)

# also works when no closures are applied
o <- daedalus(
  "CAN", "influenza_2009",
  "economic_closures", "high",
  time_end = 30
)
get_fiscal_costs(o)
}
