% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{get_value_school_year}
\alias{get_value_school_year}
\title{Calculate the value of a school year}
\usage{
get_value_school_year(gni)
}
\arguments{
\item{gni}{The GNI per capita of a country. Must be a single value. See
\link[daedalus.data:life_value]{daedalus.data::country_gni} for values.}
}
\value{
A single value giving the value of a school year given the country
GNI per capita and an expected work period of 45 years.
}
\description{
Calculate the value of a school year
}
\keyword{internal}
