% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_value_data.R
\docType{data}
\name{life_value}
\alias{life_value}
\alias{country_gni}
\title{Values of statistical lives lost}
\format{
\subsection{\code{life_value}}{

A list with 67 elements, each corresponding to a country or territory, and
giving the age-group-specific value of a statistical life lost. Each list
element is a numeric vector of four values giving the value per age group,
in the following order: 0-4, 5-19, 20-65, 65+.
}

\subsection{\code{country_gni}}{

A list with 67 elements, each corresponding to a country or territory, and
giving the GNI (gross national income) in international dollars.
}
}
\source{
Multiple sources; see processing details in `data-raw/life_value.R.

Multiple sources; see processing details in `data-raw/life_value.R.
}
\usage{
life_value

country_gni
}
\description{
Values in dollars (scaled by purchasing power parity) of a
statistical life lost in each of the four DAEDALUS age groups, accounting
for life expectancy.
}
\keyword{datasets}
