% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helpers.R
\name{make_initial_state}
\alias{make_initial_state}
\title{Generate a default initial state for DAEDALUS}
\usage{
make_initial_state(country, initial_state_manual)
}
\arguments{
\item{country}{A country or territory object of class \verb{<daedalus_country>},
\strong{or} a country or territory name from those included in the package;
see \link[daedalus.data:country_names_codes]{daedalus.data::country_names}, \strong{or} a country ISO2 or ISO3 code; see
\link[daedalus.data:country_names_codes]{daedalus.data::country_codes_iso2c} and
\link[daedalus.data:country_names_codes]{daedalus.data::country_codes_iso3c}.
Country-specific data such as the community and workplace contacts, the
demography, and the distribution of the workforce into economic sectors is
automatically accessed from package data for the relevant country name if it
is passed as a string.
To override package defaults for country characteristics, pass a
\verb{<daedalus_country>} object instead. See \code{\link[=daedalus_country]{daedalus_country()}} for more.}

\item{initial_state_manual}{An optional \strong{named} list with the names
\code{p_infectious}, \code{p_asymptomatic}, and \code{p_immune}.
\code{p_infectious} and \code{p_asymptomatic} give the proportion of
infectious and symptomatic individuals in each age group and economic sector.
Defaults to \code{1e-6} and \code{0.0} respectively.
\code{p_immune} may be a single number in the range \verb{0.0 <= p_immune <= 1.0} or a
4-element vector in that range (the number of age groups in the model), for
the proportion of individuals in the population or in each age group that
have some pre-existing immunity to infection (reduced susceptibility). See
\strong{Details} for more.}
}
\value{
An array with as many dimensions as \code{N_VACCINE_DATA_GROUPS}
(currently, 3); rows specify the age and economic groups, columns specify the
epidemiological compartments (including new infections and hospitalisations),
and array layers hold information on vaccination status (including new
vaccinations).
}
\description{
Function to prepare the model initial state. Assumes that
1 in every million individuals is initially infected, and that 60\% are
asymptomatic infections. This does not affect the actual probability of
asymptomatic infections in the simulation, which is a property of a
\verb{<daedalus_infection>}.
}
\keyword{internal}
