% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_initial_state}
\alias{make_initial_state}
\title{Generate a default initial state for DAEDALUS}
\usage{
make_initial_state(country, initial_state_manual)
}
\arguments{
\item{country}{A country or territory object of class \verb{<country>}.
Available country and territory names are given in \link{country_names}.
Country-specific data such as the community and workplace contacts, the
demography, and the distribution of the workforce into economic sectors is
automatically accessed from package data for the relevant country name.
To override package defaults for contact data values, use the \code{\link[=set_data]{set_data()}}
method for the \verb{<country>} class. See \code{\link[=country]{country()}} for more.}

\item{initial_state_manual}{An optional \strong{named} list with the names
\code{p_infectious} and \code{p_asymptomatic} for the proportion of infectious and
symptomatic individuals in each age group and economic sector.
Defaults to \code{1e-6} and \code{0.0} respectively.}
}
\value{
An array with as many dimensions as \code{N_ECON_STRATA} (currently, 46)
with each layer giving the proportion of individuals of each group in each
epidemiological compartment.
}
\description{
Function to prepare the model initial state.
}
\keyword{internal}
