% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{make_model_parameters}
\alias{make_model_parameters}
\alias{make_infection_parameters}
\title{Make infection parameters list for DAEDALUS}
\usage{
make_infection_parameters(epidemic, infect_params_manual)
}
\description{
Make infection parameters list for DAEDALUS
}
\keyword{internal}
