% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{make_parameters}
\alias{make_parameters}
\title{Make a parameter list for DAEDALUS}
\usage{
make_parameters(country, ...)
}
\arguments{
\item{country}{A string giving a country name for which stored demographic
and economic data are extracted.}

\item{...}{User-specified values to replace default pathogen parameter values
.}
}
\description{
Helper function to prepare a parameter list for \code{daedalus()},
allowing the use of country-specific contact matrices and user-specified
pathogen parameter values.
}
