% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_vaccination.R
\name{make_susc_matrix}
\alias{make_susc_matrix}
\title{Prepare susceptibility matrix for a vaccine-country pair}
\usage{
make_susc_matrix(vaccination, country)
}
\arguments{
\item{country}{A country or territory object of class \verb{<daedalus_country>},
\strong{or} a country or territory name from those included in the package;
see \link[daedalus.data:country_names_codes]{daedalus.data::country_names}, \strong{or} a country ISO2 or ISO3 code; see
\link[daedalus.data:country_names_codes]{daedalus.data::country_codes_iso2c} and
\link[daedalus.data:country_names_codes]{daedalus.data::country_codes_iso3c}.
Country-specific data such as the community and workplace contacts, the
demography, and the distribution of the workforce into economic sectors is
automatically accessed from package data for the relevant country name if it
is passed as a string.
To override package defaults for country characteristics, pass a
\verb{<daedalus_country>} object instead. See \code{\link[=daedalus_country]{daedalus_country()}} for more.}
}
\value{
A matrix with dimensions as follows:
\itemize{
\item Rows: number of age and economic groups in \code{country};
\item Cols: number of vaccination strata in the DAEDALUS model, given as
\code{N_VAX_STRATA}.
}
}
\description{
Helper function to prepare a susceptibility matrix to be used internally to
modulate the number of infections in vaccinated groups.
}
\keyword{internal}
