% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{model_constants}
\alias{model_constants}
\alias{N_AGE_GROUPS}
\alias{N_VACCINE_STRATA}
\alias{AGE_GROUPS}
\title{DAEDALUS model constants}
\format{
An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{character} of length 4.
}
\usage{
N_AGE_GROUPS

N_VACCINE_STRATA

AGE_GROUPS
}
\value{
Values for the model constants.
}
\description{
Frequently used values for the DAEDALUS model related
to the model population and structure. See also \link{epi_constants} for constants
specific to the epidemiological model.
}
\details{
The DAEDALUS model requires the following values to be fixed.
\itemize{
\item Number of age groups: 4
\item Number of vaccination strata: 1
\item Age group bins: 0-4 years, 5-19 years, 20-65 years (working age), 65+ years
}
}
\keyword{constant}
\keyword{datasets}
\keyword{internal}
