% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{model_constants}
\alias{model_constants}
\alias{N_AGE_GROUPS}
\alias{N_VACCINE_STRATA}
\alias{AGE_GROUPS}
\alias{i_WORKING_AGE}
\alias{N_ECON_SECTORS}
\alias{i_EDUCATION_SECTOR}
\alias{N_ECON_STRATA}
\alias{i_NOT_WORKING}
\alias{DIM_AGE_GROUPS}
\alias{DIM_EPI_COMPARTMENTS}
\alias{DIM_ECON_SECTORS}
\alias{N_OUTPUT_COLS}
\title{DAEDALUS model constants}
\format{
An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{character} of length 4.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.
}
\usage{
N_AGE_GROUPS

N_VACCINE_STRATA

AGE_GROUPS

i_WORKING_AGE

N_ECON_SECTORS

i_EDUCATION_SECTOR

N_ECON_STRATA

i_NOT_WORKING

DIM_AGE_GROUPS

DIM_EPI_COMPARTMENTS

DIM_ECON_SECTORS

N_OUTPUT_COLS
}
\value{
Values for the model constants.
}
\description{
Frequently used values for the DAEDALUS model related
to the model population and structure. See also \link{epi_constants} for constants
specific to the epidemiological model.
}
\details{
The DAEDALUS model requires the following values to be fixed.
\itemize{
\item Number of age groups: 4
\item Number of vaccination strata: 1
\item Age group bins: 0-4 years, 5-19 years, 20-65 years (working age), 65+ years
\item Index of the working-age age groups: 3; see \code{AGE_GROUPS}
\item Number of economic sectors: 45
\item Index of individuals not working as a layer in the 3D state tensor: 1
\item Number of economic strata: 46; note that this is always one more than the
number of economic sectors, with the additional stratum for those not in work
which includes all individuals not of working age, and a proportion of
working age individuals.
\item Array dimension of age groups: 1
\item Array dimension of epidemiological compartments: 2
\item Array dimension of economic sectors: 3
}
}
\keyword{model_constant}
