% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus_rtm.R
\name{prepare_contacts}
\alias{prepare_contacts}
\alias{make_conmat_large}
\alias{make_work_contacts}
\alias{make_consumer_contacts}
\title{Make large contact matrix for Cpp model.}
\usage{
make_conmat_large(country)

make_work_contacts(country)

make_consumer_contacts(country)
}
\arguments{
\item{country}{A \verb{<daedalus_country>}.}
}
\value{
\enumerate{
\item \code{make_conmat_large()} returns a 49x49 contact matrix scaled by the size of
demography groups.
\item \code{make_work_contacts()} returns a 45-element vector (for the number of
economic sectors) scaled by the number of workers per sector.
\item \code{make_consumer_contacts()} returns a 45x4 contact matrix with each row
scaled by the number of workers per sector. Dimensions are the number of
economic sectors and the number of age groups.
}
}
\description{
Make large contact matrix for Cpp model.
}
\keyword{internal}
