% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_output.R
\name{prepare_output}
\alias{prepare_output}
\title{Prepare DAEDALUS data}
\usage{
prepare_output(output, country)
}
\arguments{
\item{output}{Output from \code{\link[=daedalus_internal]{daedalus_internal()}}, expected to be a list.}

\item{country}{A \verb{<daedalus_country>} object from which to get data on the
number of demographic, economic, and vaccine groups.}
}
\value{
A \verb{<data.frame>} in long or 'tidy' format with the columns
"time", "age_group", "compartment", "econ_sector", and "value", for the
age-group specific value of the number of individuals in each economic sector
and epidemiological compartment in each timestep.
}
\description{
Convert DAEDALUS data into a long-format \verb{<data.frame>}.
}
\details{
There are 45 economic sectors, given by \code{N_ECON_SECTORS}, with the first
'sector' representing non-working individuals. All age groups that are
considered non-working are in this first sector (children and retirees),
while working-age individuals may be distributed flexibly into the various
economic sectors (including non-working).
}
\keyword{internal}
