% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_output.R
\name{prepare_output}
\alias{prepare_output}
\title{Prepare DAEDALUS data}
\usage{
prepare_output(output)
}
\arguments{
\item{output}{Output from \code{\link[=daedalus]{daedalus()}} of the class \verb{<deSolve>}.}
}
\value{
A \verb{<data.frame>} in long or 'tidy' format with the columns
"time", "age_group", "compartment", and "value", for the age-group specific
value of the number of individuals in each epidemiological compartment in
each timestep.
}
\description{
Convert DAEDALUS data into a long-format \verb{<data.frame>}.
}
