% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_country.R, R/pkg_generics.R
\name{prepare_parameters}
\alias{prepare_parameters}
\alias{prepare_parameters.country}
\title{Prepare country parameters for model}
\usage{
\method{prepare_parameters}{country}(x)

prepare_parameters(x, ...)
}
\arguments{
\item{x}{An S3 object with an appropriate method.}

\item{...}{Not used; included for compatibility with methods.}
}
\value{
A list of parameters suitable for the DAEDALUS model.
\code{\link[=prepare_parameters.country]{prepare_parameters.country()}} is the only method, and returns the country
parameters.
}
\description{
Generic for the \link{prepare_parameters} methods associated with
\pkg{daedalus} classes.
}
\details{
\subsection{Country parameters}{

Country contact data is processed as follows:
\itemize{
\item \code{contact_matrix}: scaled by its leading eigenvalue, and with each column
\code{j} scaled by the \code{j}-th element of the country demography vector (i.e.,
scaling contacts from each age group by the size of that group).
}

The returned parameter list consists of:
\itemize{
\item \code{demography}: the demography vector;
\item \code{contact_matrix}: the scaled contact matrix suitable for multiplication
with \eqn{R_0} in force of infection calculations;
\item \code{contacts_workplace}: the contacts in workplaces scaled by their largest
value (which is the leading eigenvalue of the diagonal matrix of contacts);
\item \code{contacts_consumer_worker}: contacts in workplaces distributed in
proportion to the demography distribution, and scaled by the largest singular
value (similar to eigenvalue for non-square matrices).
\item \code{contacts_between_sectors}: dummy matrix of zeros for future use.
}
}
}
\keyword{internal}
