% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reff_calculation.R
\name{reff_calculation}
\alias{reff_calculation}
\alias{r_eff}
\title{Calculate the effective R}
\usage{
r_eff(r0, state, cm)
}
\arguments{
\item{r0}{The basic reproduction number.}

\item{state}{The ODE state variable. Must be an array.}

\item{cm}{The contact matrix.}
}
\value{
A single numeric value for the \eqn{R_\text{eff}}.
}
\description{
A simple calculation for the effective \eqn{R_\text{eff}}, allowing for
heterogeneity in social contacts and the proportion of each age group
remaining susceptible.
This assumes a single susceptible stratum with full susceptibility, and will
be updated in future versions to account for the addition of vaccination
strata with reduced susceptibility.
}
\keyword{internal}
