% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_infection.R, R/class_vaccination.R,
%   R/pkg_generics.R
\name{set_data}
\alias{set_data}
\alias{set_data.daedalus_infection}
\alias{set_data.daedalus_vaccination}
\title{Set parameters in DAEDALUS classes}
\usage{
\method{set_data}{daedalus_infection}(x, ...)

\method{set_data}{daedalus_vaccination}(x, ...)

set_data(x, ...)
}
\arguments{
\item{x}{An S3 class object from the \pkg{daedalus} package of the
\verb{<daedalus_country>} or \verb{<infection>} class.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Named optional arguments for
parameters to be changed, with their new values. The only values allowed for
\verb{<daedalus_country>} objects are "contact_matrix", "contacts_workplace", and
"contacts_consumer_worker".}
}
\value{
An S3 object of the same class as input \code{x}.
}
\description{
Generic and methods for S3 classes for safely setting class
parameters. Only parameters considered safe to change -- mostly contact data
in the \verb{<country>}, but all parameters in \verb{<infection>} -- can be changed in
this way.
}
\examples{
# simple example of setting all contacts to 1
country_A <- daedalus_country("United Kingdom")
country_A

country_A <- set_data(country_A, contact_matrix = matrix(1, 4, 4))
country_A

disease_x <- daedalus_infection("sars_cov_1")
disease_x <- set_data(disease_x, r0 = 3.0)
disease_x
}
\keyword{generics}
