% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_country.R, R/pkg_generics.R
\name{set_data}
\alias{set_data}
\alias{set_data.country}
\title{Set \verb{<country>} parameters}
\usage{
\method{set_data}{country}(x, ...)

set_data(x, ...)
}
\arguments{
\item{x}{An S3 class object from the \pkg{daedalus} package. Currently only
\verb{<country>} objects are supported.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Named optional arguments for
parameters to be changed, with their new values. The only values allowed for
\verb{<country>} objects are "contact_matrix", "contacts_workplace", and
"contacts_consumer_worker".}
}
\value{
An S3 object of the same class as input \code{x}.
}
\description{
Generic and methods for S3 classes for safely setting class
parameters. Only parameters considered safe to change -- mostly contact data
-- can be changed in this way.
}
\examples{
# simple example of setting all contacts to 1
country_A <- country("United Kingdom")
country_A

country_A <- set_data(country_A, contact_matrix = matrix(1, 4, 4))
country_A
}
