% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{summary_constants}
\alias{summary_constants}
\alias{SUMMARY_MEASURES}
\alias{SUMMARY_GROUPS}
\title{Model output measures and groups for summaries and derived time-series}
\format{
An object of class \code{character} of length 3.

An object of class \code{character} of length 2.
}
\usage{
SUMMARY_MEASURES

SUMMARY_GROUPS
}
\description{
Model output measures and groups for summaries and derived time-series
}
\keyword{datasets}
\keyword{summary_constants}
