% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_vaccination.R
\name{uptake_percent_to_number}
\alias{uptake_percent_to_number}
\title{Convert an uptake limit from a percentage to a number}
\usage{
uptake_percent_to_number(uptake_limit, country)
}
\arguments{
\item{uptake_limit}{A single number giving the upper limit for the
\emph{percentage} of the population that can be vaccinated. When this limit is
reached, vaccination ends.}

\item{country}{A \verb{<daedalus_country>} object or a 2- or 3-character string
that can be coerced to a \verb{<daedalus_country>} (e.g. \code{"GBR"} for the United
Kingdom). Used to determine when vaccination should end.}
}
\value{
A single number giving the total number of individuals expected to
take up vaccination. If \code{country} is \code{NULL}, the function returns \code{NULL}.
}
\description{
Convert an uptake limit from a percentage to a number
}
\keyword{internal}
