% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_npi.R
\name{validate_npi_input}
\alias{validate_npi_input}
\title{Validate npi inputs}
\usage{
validate_npi_input(x, country, infection, response_time, max_duration = 365)
}
\arguments{
\item{x}{An object to be validated as input to the \code{vaccine_investment}
argument of \code{\link[=daedalus]{daedalus()}}.}

\item{country}{A country or territory object of class \verb{<daedalus_country>},
\strong{or} a country or territory name from those included in the package;
see \link[daedalus.data:country_names_codes]{daedalus.data::country_names}, \strong{or} a country ISO2 or ISO3 code; see
\link[daedalus.data:country_names_codes]{daedalus.data::country_codes_iso2c} and
\link[daedalus.data:country_names_codes]{daedalus.data::country_codes_iso3c}.
Country-specific data such as the community and workplace contacts, the
demography, and the distribution of the workforce into economic sectors is
automatically accessed from package data for the relevant country name if it
is passed as a string.
To override package defaults for country characteristics, pass a
\verb{<daedalus_country>} object instead. See \code{\link[=daedalus_country]{daedalus_country()}} for more.}

\item{infection}{An infection parameter object of the class
\verb{<daedalus_infection>}, \strong{or} an epidemic name for which data are provided
in the package; see \link[daedalus.data:epidemic_data]{daedalus.data::epidemic_names} for historical epidemics
or epidemic waves for which parameters are available.
Passing the name as a string automatically accesses the default parameters
of an infection. Create and pass a \verb{<daedalus_infection>} to tweak infection
parameters using \code{\link[=daedalus_infection]{daedalus_infection()}}.}

\item{response_time}{A single numeric value for the time in days
at which the selected response is activated. This is ignored if the response
has already been activated by the hospitalisation threshold being reached.
Defaults to 30 days. Responses have a default maximum duration of 365 days.
This can be changed by passing a \verb{<daedalus_npi>} object to
\code{response_strategy}.}
}
\value{
A \verb{<daedalus_npi>} object.
}
\description{
Validate npi inputs
}
\keyword{internal}
