% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_vaccination.R
\name{validate_vaccination_input}
\alias{validate_vaccination_input}
\title{Validate vaccination inputs}
\usage{
validate_vaccination_input(x)
}
\arguments{
\item{x}{An object to be validated as input to the \code{vaccine_investment}
argument of \code{\link[=daedalus]{daedalus()}}.}
}
\value{
A \verb{<daedalus_vaccination>} object.
}
\description{
Validate vaccination inputs
}
\keyword{internal}
